### 2024-06-05 #6847 @editorialbot
[PRE REVIEW]: pyforce: Python Framework for data-driven model Order Reduction of multi-physiCs problEms
**Submitting author:** @Steriva (<a href="http://orcid.org/0000-0001-9997-4101">Stefano Riva</a>)
**Repository:** https://github.com/ERMETE-Lab/ROSE-pyforce
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
