### 2024-06-05 #6847 @danielskatz
👋 @Steriva - thanks for your submission.

Please add a statement of need section to your paper, as in our [example paper](https://joss.readthedocs.io/en/latest/example_paper.html).

In addition, please fix the invalid DOIs that editorialbot found. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
