### 2024-06-05 #6847 @danielskatz
Please also add countries to all affiliations.
