### 2024-06-05 198b05c
Update JOSS paper
