+++ joss-paper/paper.bib
@@ -5,7 +5,7 @@ volume = {421},
 pages = {113105},
 year = {2024},
 issn = {0029-5493},
-doi = {https://doi.org/10.1016/j.nucengdes.2024.113105},
+doi = {10.1016/j.nucengdes.2024.113105},
 url = {https://www.sciencedirect.com/science/article/pii/S002954932400205X},
 author = {Antonio Cammi and Stefano Riva and Carolina Introini and Lorenzo Loi and Enrico Padovani},
 keywords = {Hybrid Data-Assimilation, Generalized Empirical Interpolation Method, Indirect Reconstruction, Sensors positioning, Molten Salt Fast Reactor, Noisy data},
@@ -34,7 +34,8 @@ abstract = {Sensor positioning and real-time estimation of non-observable fields
       year={2024},
       eprint={2401.07300},
       archivePrefix={arXiv},
-      primaryClass={math.NA}
+      primaryClass={math.NA},
+	  doi = {10.48550/arXiv.2401.07300}
 }
 
 @misc{BarattaEtal2023,
@@ -45,23 +46,24 @@ abstract = {Sensor positioning and real-time estimation of non-observable fields
   howpublished = {preprint}
 }
 
-@book{quarteroni2015reduced,
-author = {Quarteroni, A and Manzoni, A and Negri, F},
-isbn = {9783319154312},
-publisher = {Springer International Publishing},
-series = {UNITEXT},
-title = {{Reduced Basis Methods for Partial Differential Equations: An Introduction}},
-year = {2015},
-url = {https://link.springer.com/book/10.1007/978-3-319-15431-2}
+@book{rozza_model_2020,
+	title = {Model {Order} {Reduction}: {Volume} 2: {Snapshot}-{Based} {Methods} and {Algorithms}},
+	isbn = {978-3-11-067149-0},
+	url = {https://doi.org/10.1515/9783110671490},
+	publisher = {De Gruyter},
+	author = {Rozza, Gianluigi and Hess, Martin and Stabile, Giovanni and Tezzele, Marco and et al.},
+	year = {2020},
+	doi = {10.1515/9783110671490},
 }
 
-@inproceedings{demo_complete_2019,
-	title = {A complete data-driven framework for the efficient solution of parametric shape design and optimisation in naval engineering problems},
-	booktitle = {{MARINE} 2019: {VIII} {International} {Conference} on {Computational} {Methods} in {Marine} {Engineering}},
-	author = {Demo, Nicola and Tezzele, Marco and Mola, Andrea and Rozza, Gianluigi},
-	month = may,
-        pages = {1-12},
-	year = {2019},
+@misc{demo2019complete,
+      title={A complete data-driven framework for the efficient solution of parametric shape design and optimisation in naval engineering problems}, 
+      author={Nicola Demo and Marco Tezzele and Andrea Mola and Gianluigi Rozza},
+      year={2019},
+      eprint={1905.05982},
+      archivePrefix={arXiv},
+      primaryClass={math.NA},
+	  doi = {10.48550/arXiv.1905.05982}
 }
 
 @article{maday_generalized_2015,
@@ -84,7 +86,7 @@ url = {https://link.springer.com/book/10.1007/978-3-319-15431-2}
 	copyright = {All rights reserved},
 	issn = {0045-7825},
 	url = {https://www.sciencedirect.com/science/article/pii/S0045782522007290},
-	doi = {https://doi.org/10.1016/j.cma.2022.115773},
+	doi = {10.1016/j.cma.2022.115773},
[…]
