+++ joss-paper/paper.bib
@@ -56,7 +56,7 @@ abstract = {Sensor positioning and real-time estimation of non-observable fields
 	doi = {10.1515/9783110671490},
 }
 
-@misc{demo2019complete,
+@misc{demo_complete_2019,
       title={A complete data-driven framework for the efficient solution of parametric shape design and optimisation in naval engineering problems}, 
       author={Nicola Demo and Marco Tezzele and Andrea Mola and Gianluigi Rozza},
       year={2019},
