### 2024-06-14 1bc1d76
Add EIM offline algorithm for sensor placement
