+++ pyforce/pyforce/offline/eim.py
@@ -1,0 +1,216 @@
+# Offline Phase: Empirical Interpolation Method
+# Author: Stefano Riva, PhD Student, NRG, Politecnico di Milano
+# Latest Code Update: 14 June 2024
+# Latest Doc  Update: 14 June 2024
+
+import numpy as np
+from scipy import linalg
+
+from dolfinx.fem import (Function, FunctionSpace)
+
+from pyforce.tools.backends import norms, LoopProgress
+from pyforce.tools.functions_list import FunctionsList
+    
+# EIM: offline
+class EIM():
+    r"""
+    This class is used to perform the offline phase of the Empirical Interpolation Method (EIM) to a scalar field.
+    Given a list of training snapshots, this class generates the magic functions and points through a greedy algorithm.
+
+    Parameters
+    ----------
+    mesh : np.ndarray
+        Mesh Points shaped as :math:`\mathcal{N}_h\times gdim`, given :math:`gdim` be the number of independent coordinates (1,2 or 3).
+    name : str
+        Name of the snapshots (e.g., temperature T)
+    
+    """
+    def __init__(self, mesh: np.ndarray, name: str) -> None:
+        
+        self.mesh = mesh
+        self.name = name
+        
+        self.Nh = mesh.shape[0]
+        self.gdim = mesh.shape[1]
+        
+    def offline(self, train_snap: FunctionsList, Mmax: int, _xm = None, verbose = False):
+        r"""
+        The greedy algorithm chooses the magic functions and magic points by minimising the reconstruction error.
+        
+        Parameters
+        ----------
+        train_snap : FunctionsList
+            List of snapshots serving as training set.
+        Mmax : int
+            Integer input indicating the maximum number of functions and sensors to define
+        _xm : list, optional (default = None)
+            User-defined available indices positions for the points (sensors), if `None` the positions are all mesh elements.
+        verbose : boolean, optional (Default = False) 
+            If `True`, print of the progress is enabled.
+            
+        Returns
+        ----------
+        maxAbsErr : np.ndarray
+            Maximum absolute error measured in :math:`L^\infty` (max norm)
+        maxRelErr : np.ndarray
+            Maximum relative error measured in :math:`L^\infty` (max norm)
+        beta_coeff : np.ndarray
+            Matrix of the reduced coefficients :math:`\{ \beta_m \}_{m=1}^M`, obtained by greedy procedure
+        """
+        self.Ns = len(train_snap)
+
+        # Generate sensor library
+        if _xm is None:
+            xm = np.arange(0, self.mesh.shape[0], 1, dtype=int)
+        else:
+            assert max(_xm) < self.mesh.shape[0], "Indices are out of range"
+            assert min(_xm) > 0, "Indices are out of range"
+            assert len(_xm) < self.mesh.shape[0], "List of positions not compatible with the mesh points"
+            
+            xm = np.asarray(_xm, dtype=int)
+            
+        beta_coeff = np.zeros((len(train_snap), Mmax))
+        
+        self.magic_fun = FunctionsList(dofs = train_snap.fun_shape)
+        self.generating_fun = list()
+        self.magic_points = {'idx': list(), 'points': list()}
+        
+        maxAbsErr = list()
+        maxRelErr = list()
+        
+        snap_matrix = train_snap.return_matrix()
+        
+        # Generating the first magic function and associated point
+        mm = 0
+        self.generating_fun.append( int(np.argmax(np.max(abs(snap_matrix[xm]), axis=0))) )
[…]
