### 2024-07-02 #6847 @alberto743
> Hi @alberto743, thanks for your prompt response. Your review timing after July 10th works for us. I am happy to add you to our reviewer list. Please confirm if this works for you.

Ok, fine for me.
