### 2024-07-24 ERMETE-Lab/ROSE-pyforce#6 @alberto743
License file with incomplete authors
License is MIT, but the authors provided in the heading of the license file is inconsistent.
I guess that instead of "Copyright (c) 2024 ROSE-pyforce", there should be something like "Copyright (c) 2024 Politecnico di Milano, Italy".
