### 2024-07-24 ERMETE-Lab/ROSE-pyforce#6 @alberto743
It should be noted that this information is later available in the `__init__.py` file and in the heading of each source.
