### 2024-07-24 #6950 @alberto743
Fix license text: https://github.com/ERMETE-Lab/ROSE-pyforce/issues/6
