### 2024-07-26 ERMETE-Lab/ROSE-pyforce#6 @Steriva
I have updated the license, after having discussed with the other contributors from Polimi.
