### 2024-07-26 #6950 @Steriva
License text has been fixed with [https://github.com/ERMETE-Lab/ROSE-pyforce/pull/7](pull request #7)
