### 2024-08-09 ERMETE-Lab/ROSE-pyforce#8 @damar-wicaksono
JOSS Submission Review (Reviewer 2): Documentation
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6950).

I'm currently reviewing the documentation and below are some remarks, questions, and suggestions that the authors may want to consider regarding the documentation (and testing); please let me know if clarifications are needed. What listed below are partial as I haven't gone through to the whole documentation. I will put additional remarks/suggestions if/when they are needed.

### Landing page

- [x] I'd suggest that the landing page of the documentation (the one reached by clicking the following [link](https://ermete-lab.github.io/ROSE-pyforce/index.html) is directly linked to [intro](https://ermete-lab.github.io/ROSE-pyforce/intro.html). That is, put some contents of `intro.md` inside `index.rst` (i.e., description and how to cite); the table of contents may appear here. The introduction section may then contain the description on what is ROM? and the Data-Driven ROM techniques.

PS: Clicking the `docs` badge in the GitHub repository actually brings us to the intro page anyway and not `index.html`.

### Welcome to pyforce's documentation

**Description**

-  [x] _The package is part of the **ROSE (Reduced Order modelling with data-driven techniques for multi-phySics problEms)**_

What is actually "ROSE", is it a project? a software framework? or something else? Perhaps there is a link to the project site?

- [x] As a suggestion, I think there is no need to repeat the preposition "_at_" multiple times in: "the package...is aimed at reducing..., at searching..., and at integrating" because they are already parallel
- [x] Perhaps revise "either regularised with Tikhonov or not" to "either with or without Tikhonov's regularisation"
- [x] "_not only restricted in the Nuclear Engineering world._": perhaps replace it with "_not restricted to the Nuclear Engineering world._"
- [x] "_the API documentation and some examples **on** how to use the various modules of the package._": perhaps replace **on** with **of**

**How to cite pyforce**

- [x] "_If you are going to use *pyforce* in your research work, please cite the following articles. The authors would be pleased if you could cite the relevant papers:_": I think one of these two sentences is redundant, pick either one or combine them. Also, if one is simply _going to use_ pyforce (they may in the end not use it), perhaps there is no need to ask for a citation.

**What is ROM**

- [x] As noted above, perhaps move this and the subsection below it to a new page
- [x] "_In scientific literature,..._": missing a comma
- [x] "_Full Order (FOM) Model_": Perhaps "Full Order Model (FOM)"
- [x] "_...with the developments in data-driven modelling,..._": missing a comma
- [x] "_...a lot of interest ... have been raised..._": I might be wrong but I think "(a lot of) interest" is uncountable so it should have been singular ("has" instead of "have")

**Reduced Basis Methods**

- [x] "_Among all ROM methods, Reduced Basis (RB)... are a well-established..._": Maybe "Reduced Basis (RB) methods" instead of just "Reduced Basis (RB)"
- [x] "_...a set of RB functions are derived from an ensemble of high-fidelity solutions..._": "a set of..." is singular so use "is" instead of "are"
- [x] "_...and the **POD**..._": Perhaps expand the term POD here (just like RB)
- [x] There is a missing semicolon separating the spatial variable $\mathbf{x}$ and $\mu$ in the equation
- [x] "_This project mainly focuses on..._": Perhaps directly use "pyforce" instead of "this project"

**Data-Driven ROM techniques**

- [x] "_...(DDROM) are a set of techniques..._": I think DDROM should be singular so "is" instead of "are"
- [x] "_...techniques, which combines theoretical modelling with real data collecting from a physical system..._": I believe the which clause is restrictive (it defines DDROM as a set of techniques) so instead of "which" with a comma, use "that" without a comma.
- [x] "_...the measure of the temperature in a pipe..._": Perhaps this should be "the measured temperature in a pipe"
- [x] "_...by searching for the dominant physics into the system._": Replace "into" with "in"
- [x] "_This process of encoding can be done through POD, GEIM,..._": I assume GEIM is "Generalised Empirical Interpolation Method" but perhaps it should be expanded here.
- [x] "_...and in general,..._": missing a comma
- [x] "This process of encoding...represents the heaviest part from the computational point of view.": What is the scope of this sentence? I would assume getting the snapshots is computationally more expensive, once the snapshots have been obtained, carrying out POD or GEIM should be less expensive, or?
- [x] "_Then, the reduced coordinate system is used to select the most important locations in the domain to place experimental sensors._": I find this sentence is a bit awkward, perhaps: "Then, the reduced coordinate system is used to **identify** the most important locations in the domains **where** experimental sensors **can be placed**."
- [x] "_...so that an improved system of coordinate**s**..._": missing an "s"

### Installation notes

- [x] "_The package has been tested on MacOS and Linux machines..._": replace "the package" with "pyforce".
- [x] I think the dependencies should be listed along with the version numbers (at least the one the authors used).
- [x] Instead of showing how to set up a conda environment step-by-step in this section of the documentation, the authors should directly refer to the available `yaml` file used to set up a conda environment just like the instruction in the README.
- [x] Also provide a note stating that currently pyforce can only be obtained by directly cloning the repository (not in PyPI or conda repository)
- [x] "_*docs/* folder can be generate with_": Why such a folder need to be generated?
- [x] Regarding the documentation, perhaps list the dependencies first before giving the instruction on how to build the docs

## Tutorials

The tutorials are provided in great detail and appreciated. I would propose to pick one of the tutorials to serve as a "Getting Started" part of the documentation that serves as an entry point in using the package. This "Getting Started" would provide more details regarding the features of the package gradually with enough detail. I assume the DFG2 benchmark may serve as that entry point and my remarks below reflect that.

- [x] "_in the tutorials, alternatevely the data can be download from [Zenodo](https://doi.org/10.5281/zenodo.11483677)._": typo "alternatevely"
- [x] In the tutorial, the authors repeatedly use the idiom "at first" which may be inappropriate as it is used more in the context of contrasting something later on ("at first it's A, but in the end it's B"). They appear numerous times, so please double check them.

## [Unsteady Laminar Navier Stokes - DFG2 benchmark — pyforce 0.1.0 documentation](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_DFG2_benchmark.html)

- [x] I think this page should include explicitly the purpose of this tutorial, what is the expected outcome of carrying out this exercise. Currently it contains only the description.

## Generation of Snapshots

- [x] Should the title be "Generation of Snapshots" or "DFG2 Benchmark: Generation of Snapshots" instead of "Flow past a cylinder (DFG 2D-3 benchmark)"
- [x] In general, what level of familiarity the users need to have with FEniCS to carry out this tutorial? When applicable perhaps provide links to the relevant sections of documentation of FEniCS/DOLFIN
- [x] "_The kinematic velocity_": Shouldn't this be "kinematic viscosity"?

**Assembling variational forms**

- [x] "*EI* = Euler Implitict": Typo "Implitict"
- [x] Please add a short description on what is the module `ns.py` provided by the authors (I may guess that it's to solve NS equation but because it's coming in a single file it should be explained).
- [x] I would avoid using `*` to import everything from `ns` as it obscures what functions coming from which package/module.
- [x] BDF2 is selected via `time_adv = 'BDF2'` but the series of plots below show the results of the three methods.

**Setting up the tools to compute drag and lift coefficient**:

- [x] "_The FOM will be validation against..._": Perhaps "will be _validated_ against..."
- [x] "_...the FEATFLOW dataset..._": The dataset may need a reference or at least a link
- [x] "_We use `UFL` to create the relevant integrals..._": What is "UFL" in this context? I assume it's a Fenics-related term so a link would be useful.

```
params['rhoLU2'] = 0.1
get_drag_lift = drag_lift(mesh, ft, params, bound_markers['obstacle'])
```

- [x] Maybe I missed it, but was it explained somewhere about `params['rhoLU2'] = 0.1`?
- [x] I assume `drag_lift()` is coming from `ns.py` (which upon inspection is actually a class not a function; In Python, the standard practice is to name a class with a camel case instead of with underscores). It needs to be explained briefly what is this class doing as well as the important parameters for construction.

**Solving the time-dependent problem**

- [x] "_The velocity and the pressure are stored to be later saved in appropriate data structures._": I find "...stored to be later saved..." is a rather awkward construction here.
- [x] `store_snap = False` which must be `True` if the users want to follow this tutorial and move on to the next steps (offline phase)
- [x] Please state clearly in the paragraph (not in the codeblock) the assumption of the directory structure (where things should be saved). Currently, the following locations must be explicitly stated and briefly clarified:
  - `Snapshots/`
  - `./../BenchmarkData/`
  - `./../BenchmarkData/LaminarNS_DFG2/`
- [x] "_Store QoI data_": I assume QoI is quantities of interest, but this has never been referred to before

**Comparison with benchmark data**:

- [x] The line `# time_adv_schemes = ['BDF2', 'CN', 'EI']` is commented out but the plot actually shows the results from the three methods.
- [x] Is there a purpose of saving the figure: `fig.savefig('./../BenchmarkData/LaminarNS_DFG2/FOM_bench_comparison.pdf', format='pdf', dpi=300, bbox_inches='tight')`
- [x] As a suggestion: plotting specifications can get in the way of the tutorial, maybe create a helper function for plotting so users can focus on what should be plotted?

## Offline Phase

- [x] There is no need for a list if there is only one item (POD)
- [x] Is it necessary in this part to create the mesh again with `gmsh` module? `domain` is needed in the subsequent codes, but would it be possible to hide the repetition of this process so the users can just pickup from where before?
- [x] State briefly that to continue with the next tutorial steps later on all important files must be saved in a specific location

**Importing Snapshots**:

- [x] I would suggest splitting large block of codes and describe it using markdown cells instead of code comments.
- [x] Make sure to remove commented out codes so as not to confuse users
- [x] Why `u` in `var_names` is different from `U` in `stored_var_names`?
- [x] I might have done something wrong, but if I used the snapshots stored in Zenodo I obtained the following exception for this tutorial:

```
...
AssertionError: The input function dofs_fun has 2385, instead of 2356
```

The snapshots I generated locally work just fine, though.

**POD algorithm**:

- [x] "_A list will be created for each POD: 0 = p, 1 = u_": What do the numbers signify? Index?
- [x] I think $N_s$ has never been defined

```
[POD(train_snaps[field_i], var_names[field_i], use_scipy=True, verbose = True) for field_i in range(len(var_names))]
```

- [x] For the first time, briefly introduce the class `POD()`; explain it's interface especially the important inputs and expected outcome output as it is currently used here. Establish a bridge between the "theoretical" ROM and the specific constructs provided by pyforce to carry out the tasks.

**Definition of the modes**:

- [x] Split the codes of computing the modes and saving them
- [x] Programmatically, what does it mean to "define" the POD modes, explain a bit what does the method `compute_basis` actually do, which important inputs it takes and what outcome is expected once executed. Explain why particular arguments are used, i.e., why the particular `Nmax` or why `normalise=True`?
- [x] Explain why saving the modes are necessary in the context of the tutorial
- [x] State briefly the assumption of the path and filenames (`path+'/BasisFunctions'` and `path+'/BasisFunctions/basisPOD_`; and why they might be important to be able to follow the tutorial

**Computing the training error**:

- [x] "_let us compute the training errors and the POD basis coefficients._": Are the POD basis coefficients being computed by the code block below?
- [x] Briefly explain the method "train_error()` of an instance of `POD` including expected inputs and expected output.
- [x] "The max absolute and relative reconstruction error is compared for the different algorithms,...": Which different algorithms?
- [x] I think $\mathcal{P}_N[u]$ is not the reconstruction operator with $N$ basis functions but the reconstructed object; it should be $\mathcal{P}_N$ without the object being operated

**Generation of the maps**:

- [x] Maybe state explicitly that these methods are coming SciPy
- [x] Perhaps explain or say it out loud how the snapshots supposed to be sorted (by value I assume)? One can look at the codes below and guess, but it would be nice to know in advance what should happen.
- [x] Is it a linear interpolation or a cubic one?

from scipy.interpolate import interp1d

for field_i in range(len(var_names)):
        coeff_maps[field_i]['LinInt'] = [interp1d(train_pod_coeff[field_i]['X'].flatten(),
                                        train_pod_coeff[field_i]['Y'][:, rank].flatten(),
                                        kind = 'cubic', fill_value = 'extrapolate')
                                        for rank in range(Nmax)]
```

- [x] "Let us store the maps using *pickle*": Please state the purpose of saving these as pickle files (perhaps they are going to be used later on)

**Post Process: plotting the POD modes**:

- [x] There are three sets of plots; the third one is not explained. One needs to understand the local plotting function `plotModes()` to understand what is going on.
- [x] I'm having problem initially plotting the MODs in MacOS; it complains about "libgl1-mesa-glx xvfb not available" but if I remove the line `pv.start_xvfb()` everything seems to be fine after installing additional packages: `pip install trame ipywidgets` (are these listed in the dependencies?)

## Online Phase: POD-I

- [x] There is no need for a list if there is only one item (POD)
- [x] Is it necessary in this part to create the mesh again with `gmsh` module? `domain` is needed in the subsequent codes, but would it be possible to hide the repetition of this process so the users can just pickup from where before?
- [x] State briefly that to continue with the next tutorial steps later on all important files must be saved in a specific location

**Importing Snapshots and offline results**:

-  [x] Remove commented out lines

**Test error**:

- [x] `Nmax` is now `25` instead of `30`, is there any reason for that?
- [x] Introduce briefly the class `POD_project` and its role within pyforce, how can it be constructed and what the instances can do; briefly explain the corresponding method `synt_test_error()` (these are package features that should be highlighted)
- [x] Similar comment as above but for the class `PODI`
- [x] Where is this computational time coming from? I think it's not very transparent especially in relation what the output of `synt_test_error` is?

Consider the following:

```
test_errors[key][field_i][2]['CoeffEstimation']
```

Without looking deep into the documentation, for an introductory tutorial it should be explained what are the contents of `test_errors[key][field_i]`, what does the index `2` signify, and what does the key `"CoeffEstimation"` mean?

**Post Process**

- [x] `Nmax` changes again here
- [x] Consider splitting the large block of code with markdown cells to hold description/explanation
- [x] Consider creating a set of common plotting helper functions and put them in a local module (and documented) so plotting settings do not get in the way but interested users can look at the codes in detail if they want
- [x] As before, I can only execute the last part only if I remove the line: `pv.start_xvfb()` (is this an important line?)
- [x] State the purpose of the directory `path='./Online_results/'`
- [x] Explain the flag `make_video = False` and what would happen if it's `True`; it may be obvious, but it's better be explicit (e.g., "To create a video of...change the flag below to `True`")

There might be more comments to follow once I go through the whole documentation
