### 2024-08-12 ERMETE-Lab/ROSE-pyforce#9 @damar-wicaksono
JOSS Submission Review (Reviewer 2): Functionality
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6950).

The tutorials provided serve as a basis for reviewing the functionality of the software.
Overall, the tutorials can be executed successfully with additional remarks that I detail below.

The requirements based on the `environment.yml` file allows the package to be installed successfully via `conda` in MacOS 14.5 (M2).
However, when creating series of plots in the tutorials the following packages are missing:

- trame
- ipywidgets
- openpyxl
- trame-vtk

- [x] Please double check if these packages are indeed necessary (at least to follow the tutorials) and include them in the requirements (perhaps optionally).
- [x] In MacOS 14.5, the statement `pv.start_xvfb()` raises exception about `libgl1-mesa-glx xvfb not available`. However, removing this statement will allow the plots to be generated.
- [x] It appears (also in MacOS 14.5), some plotting causes a complain that `vue` must be `vue2` instead of `vue3`; this may be related to the installed version of `trame`.
- [x] The dependencies version listed in the `environment.yml` are specified exactly (i.e., specified up to the `PATCH` version) which I think may lock the versions required and can be restrictive. Is this intentional?

## Performance

As I don't see any performance claim regarding the package, I believe this criterion is irrelevant.

- [x] In one of the tutorials, the Lebesgue constant is computed by calling the function `from pyforce.offline.geim import computeLebesgue` as follows:

computeLebesgue(geim_data[var_names[field_i]][jj].magic_fun, geim_data[var_names[field_i]][jj].magic_sens)
```

Both `magic_fun` and `magic_sens` are attributes of `GEIM` instances. Is there a particular reason not to implement this function as a method of the class (`GEIM.compute_lebesgue()` or `GEIM.lebesgue()`?

- [x] I find that the tutorials are a bit cluttered with complicated plotting routines and they often get in the way of presenting the (expected) results. One straightforward way to declutter the tutorials is by extracting these routines into local modules that are only relevant for the tutorials. Alternatively, if some of the plotting routines are general in nature, they can be implemented as a high-level plotting functions directly in the code base (but only if they are general).

- [x] According to the tutorials, the method `synt_test_error()` is often called with respect to instances of classes in the online phase. The return value is a tuple but each implementation of the method have different length and for quantity that means the same the position may not be consistent (e.g., `computational_time` often appears as the last element of the tuple except for instances of the `PE()` class).

In my opinion tuple of length beyond 2 (absolute and relative errors) is already difficult to intuit which index signifies which quantity. Perhaps it would be better to store the output of the method as a dictionary or a named tuple whose relevant quantities can semantically be better accessed.

PS: What does `synt_` stand for?

- [x] High-level classes `GEIM` and `PBDW` have exactly the same name in two different modules `offline.geim` (resp. `offline.pbdw`) and `online.geim` (resp. `online.pbdw`). While this is definitely not an issue in Python, I'd argue that they can be confusing for high-level classes. We must know from where which is imported and in case we need both on, say, the same script, we have to rename one of them on the fly (via `as`). I'd propose renaming the classes such that the name indicates whether the class is related to the offline or online phase (like `POD` and `PODI`)

## Other remarks

For further discussion...

- [x] The class `POD` has a method called `GrahmSchmidt`, isn't this supposed to be `GramSchmidt`? (referring to Gram-Schmidt orthogonalization procedure?)
- [x] The convention in Python is to name a class using camel case, so name like `POD_project()` is not exactly following the convention.
- [x] I'm not an expert in DDROM, so while the tutorials are helpful in assessing the functionality of the package, I'm a bit lost at mapping the available classes both in offline and online phase and how the classes should relate to each other. Currently, it feels more like a collection of classes that are available for people to mix and match (this impression may or may not be true) and not yet as a coherent framework. It's also a bit hard thinking of these classes in a higher level (which is helpful in understanding the package more quickly) because the commonalities between them are not explicit.

Perhaps, this comment is more suitable for the documentation part but I'll put it here anyways...

The package provides the following methods/techniques:

- **Proper Orthogonal Decomposition** with Projection and Interpolation for the Online Phase
- **Generalised Empirical Interpolation Method**, either regularised with Tikhonov or not
- **Parameterised-Background Data-Weak formulation**
- an **Indirect Reconstruction** algorithm to reconstruct non-observable fields

Following the offline/online workflow, the package consists of two main sub-packages: `pyforce.offline` and `pyforce.online`. Looking at the features provided above, the following classes represent each technique:

- POD:
  - `pyforce.offline.pod.POD` implements POD in the offline phase (`pyforce.offline.pod.DiscretePOD` does not appear in the tutorial, I believe)
  - `pyforce.online.pod_interpolation.PODI` implements POD in the online phase via interpolation
  - `pyforce.online.pod_projection.POD_project` implements POD in the online phase via projection
- GEIM:
  - `pyforce.offline.geim.GEIM` implements GEIM in the offline phase
  - `pyforce.online.geim_synthetic.GEIM` implements GEIM in the online phase (without Tikhonov's regularization)
  - `pyforce.online.tr_geim_synthetic.TRGEIM` implements GEIM with Tikhonov's regularization in the online phase
- PBDW:
  - `pyforce.offline.pbdw.PBDW` implements PBDW in the offline phase
  - `pyforce.online.pbdw_synthetic.PBDW` implements PBDW in the online phase
- Sensor placement (only relevant for the offline phase):
  - `pyforce.offline.sensors.SGREEDY` (exemplified in the tutorial)
  - `pyforce.offline.sensors.gaussian_sensors` (perhaps not a main feature)
- Indirect reconstruction (only relevant for the online phase):
  - `pyforce.online.indirect_recon.PE` (exemplified in the tutorial)

It seems to me POD, GEIM, PBDW share many commonalities (the details differ, though) and I would assume that these classes serve the same overall purpose (again with their own particularities):

- [x] To what extend these methods are similar or dissimilar (in terms of the class), looking at the attributes and methods they share several similarities but because there is no parent class or abstract class (to, perhaps, mimics interface) it is hard to quickly get the picture. For example, all instances share `V` and `norm` (note that in `PBDW` it is called `norms` with an `s`) as attributes. Furthermore, the online counterparts all have the method `synt_test_error()`.
- [x] How to put the classes inside `pyforce.offline.sensors` in the context of `POD`, `GEIM`, and `PBDW`. Currently the documentation lacks the intermediate-level picture of data flow between the phases that can be translated directly to the code (the Figure on DDROM is too high-level and not readily translatable to the usage workflow and code structure).
  In the tutorials, numerous quantities are saved and loaded in between notebooks which makes it a bit hard to get the bigger picture of the package capabilities; that is, the functionalities provided by the package via its classes get a bit lost in things being saved and plotted.
- [x] Can one mix match the results of offline phase from one method with the online phase of another? Is there any restrictions?
