### 2024-08-13 ERMETE-Lab/ROSE-pyforce#10 @damar-wicaksono
JOSS Submission Review (Reviewer 2): Software paper
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6950).

Below are my remarks and suggestions regarding the software paper. Just let me know if further clarifications are need. Thank you!

- [x] I think the first paragraph of the next section ("Statement of need") should be switched as the "Summary" of the paper as it better describes the high-level functionality and purpose of the package. Furthermore, the summary should also include the target audience of the package which the authors have also stated in the last paragraph of the next section.
- [x] "_...more hostile environment compared to..._": Perhaps simply "than" is better than "compared to".
- [x] "_...and able to combine together different sources..._": I think "together" is unnecessary.
- [x] "_...and real data (i.e., measurements)..._": Perhaps to be succinct, this can be written simply as "measurement data"
- [x] "_...the dynamics of un-observable complex field..._": Somewhere else in the paper "non-observable" is used, perhaps simply pick one?

- [x] As noted above, I think the first paragraph should be switched with the paragraph in "Summary". The paragraph in Summary better describes the need that is served by the package.
- [x] Because pyforce is an abbreviation perhaps the expanded term should appear directly after "_pyforce_"
- [x] "_...mainly for the Nuclear Engineering world._": Perhaps "in" instead of "for"
- [x] "_...Nuclear Engineering..._": Is the capitalization really necessary?
- [x] "_The package is part of the ROSE..._": Is there a reference to ROSE perhaps a link or a document?
- [x] "_at reducing..., at searching... and at integrating..._": I think there is no need to repeat the preposition "at" due to already parallel structure
- [x] Figure 1 should be better clarified by the second paragraph in this section; the figure contains mathematical symbols and notations that should be defined briefly and explicitly in the passage (if these symbols and notations are kept in the end). For instance, reading the paragraph it is not clear the starting point of the whole scheme (is it the field according to the full order model, or something else?).
- [x] I would suggest that the list of implemented features are rewritten (perhaps in tabular form) to correspond to the figure better; which ones belong specifically to the offline and online phase, just the offline phase, or just the online phase. The list structure gives the impression that these features serve similar exchangeable purposes.
- [x] "_...either regularised with Tikhonov or not..._": Perhaps it would be better to write as "either with or without Tikhonov's regularisation"
- [x] "_...for the Online Phase_": Before there was no capitalization for "Online Phase". Is the capitalization intentional?

## Contribution and authorship

Currently, the paper list three authors, namely S. Riva, C. Introini, and A. Cammi.  Looking at the commit contributions [page](https://github.com/ERMETE-Lab/ROSE-pyforce/graphs/contributors), I've seen two contributors have made commits. As @Steriva is S. Riva so the question regarding "Contribution and Authorship" is settled. Is [Neko-tan](https://github.com/Neko-tan) also one of the authors?

Related to the paper, I would like to also suggest including a list of the contributions of each author separately in the paper.
Some JOSS papers (for example, [here](https://joss.theoj.org/papers/10.21105/joss.05873)) adopted the [CRediT](https://credit.niso.org/) taxonomy to clarify those roles.
Perhaps the authors would consider adding such section to the paper?
