### 2024-08-13 ERMETE-Lab/ROSE-pyforce#10 @Steriva
Thank you @damar-wicaksono for the insightful comments (including #8 and #9).
In the coming weeks, I'll update the repository following your suggestions.

Just a quick note about the authors: @Neko-tan is Carolina Introini, whereas A. Cammi is the professor supervising all the activities: I'll add the CRediT taxonomy to the paper to clarify the roles.

Thank you again!
