### 2024-08-24 4a29355
Delete pyforce/pyforce/offline/eim.py
