+++ joss-paper/paper.md
@@ -28,14 +28,18 @@ bibliography: paper.bib
 ---
 
 # Summary
-Innovative reactor technologies in the framework of Generation IV are usually characterised by harsher and more hostile environments compared to standard nuclear systems, for instance, due to the liquid nature of the fuel or the adoption of liquid salt and molten as coolant. This framework poses more challenges in the monitoring of the system itself; since placing sensors inside the reactor itself is a nearly impossible task, it is crucial to study innovative methods able to combine together different sources of information, namely mathematical models and real data (i.e., measurements) in a quick, reliable and efficient way. These methods fall into the Data-Driven Reduced Order Modelling framework, they can be very useful to learn the missing physics or the dynamics of the problem, in particular, they can be adapted to generate surrogate models able to map the out-core measurements of a simple field (e.g., neutron flux and temperature) to the dynamics of un-observable complex fields (precursors concentration and velocity). 
+*pyforce* (Python Framework for data-driven model Order Reduction of multi-physiCs problEms) is a Python library implementing Data-Driven Reduced Order Modelling (DDROM) techniques [@RMP_2024] for applications to multi-physics problems, mainly in the nuclear engineering world. These techniques have been implemented upon the dolfinx package [@BarattaEtal2023] (currently v0.6.0), part of the FEniCSx project, to handle mesh generation, integral calculation and functions storage. The package is part of the ROSE (Reduced Order modelling with data-driven techniques for multi-phySics problEms) framework, which is one of the main research topics investigated at the [ERMETE-Lab](https://github.com/ERMETE-Lab#reduced-order-modelling-with-data-driven-techniques-for-multi-physics-problems-rose-): in particular, the focus of the research activities is on mathematical algorithms aimed at reducing the complexity of multi-physics models with a focus on nuclear reactor applications, searching for optimal sensor positions and integrating experimental data to improve the knowledge on the physical systems.
 
 # Statement of need
-*pyforce* is a Python library (Python Framework for data-driven model Order Reduction of multi-physiCs problEms) implementing Data-Driven Reduced Order Modelling (DDROM) techniques [@RMP_2024] for applications to multi-physics problems, mainly for the Nuclear Engineering world. These techniques have been implemented upon the dolfinx package [@BarattaEtal2023] (currently v0.6.0), part of the FEniCSx project, to handle mesh generation, integral calculation and functions storage. The package is part of the ROSE (Reduced Order modelling with data-driven techniques for multi-phySics problEms) framework, which investigates mathematical algorithms aimed at reducing the complexity of multi-physics models with a focus on nuclear reactor applications, at searching for optimal sensor positions and at integrating experimental data to improve the knowledge on the physical systems.
-
-The techniques implemented here follow the same underlying idea expressed in \autoref{fig:darom}: in the offline (training) phase, a dimensionality reduction process retrieves a reduced coordinate system onto which the information of the mathematical model is encoded; the sensor positioning algorithm then uses this reduced set to select the optimal location of sensors according to some optimality criterion, which depends on the adopted algorithm. In the online phase, the data assimilation process begins, retrieving a novel set of reduced variables and then computing the reconstructed state through a decoding step.
[…]
