### 2024-09-03 7b9e538
Update figure description and add table for the algorithms
