+++ joss-paper/paper.md
@@ -33,26 +33,107 @@ bibliography: paper.bib
 # Statement of need
 Innovative reactor technologies in the framework of Generation IV are usually characterised by harsher and more hostile environments than standard nuclear systems, for instance, due to the liquid nature of the fuel or the adoption of liquid salt and molten as coolant. This framework poses more challenges in the monitoring of the system itself; since placing sensors inside the reactor itself is a nearly impossible task, it is crucial to study innovative methods able to combine different sources of information, namely mathematical models and measurements data (i.e., local evaluations of quantities of interest) in a quick, reliable and efficient way. These methods fall into the Data-Driven Reduced Order Modelling framework, they can be very useful to learn the missing physics or the dynamics of the problem, in particular, they can be adapted to generate surrogate models able to map the out-core measurements of a simple field (e.g., neutron flux and temperature) to the dynamics of non-observable complex fields (precursors concentration and velocity).
 
+The techniques implemented here follow the same underlying idea expressed in \autoref{fig:darom}. They all share the typical offline/online paradigm of ROM techniques: the former is computationally expensive and it is performed only once, whereas the latter is cheap from the computational point of view and allows to have quick and reliable evaluations of the state of the system by merging background model knowledge and real evaluations of quantities of interest [@maday_parameterized-background_2014].
+During the offline (also called training) phase, a *high-fidelity* or Full Order Model (FOM), usually parameterised partial differential equations, is solved several times to obtain a collections of snapshots $\mathbf{u}_{FOM}\in\mathbb{R}^{\mathcal{N}_h}$, given $\mathcal{N}_h$ the dimension of the spatial mesh, which are dependent on some parameters $\boldsymbol{\mu}_n$; then, these snapshots are used to generate a reduced representation through a set of basis functions $\{\psi_n(\mathbf{x})\}$, in this way the degrees of freedom are decresed from $\mathcal{N}_h$ to $N$, provided that $\mathcal{N}_h>>N$. This allows to approximate any solution of the FOM as follows
+
+\begin{equation}\label{eq:rb}
+u(\mathbf{x};\boldsymbol{\mu}) \simeq \sum_{n=1}^N\alpha_n(\boldsymbol{\mu})\cdot \psi_n(\mathbf{x})
+\end{equation}
+with $\alpha_n(\boldsymbol{\mu})$ as the reduced coefficients, embedding the parametric dependence. Moreover, a reduced representation allows for the search of the optimal positions of sensors in the physical domain in a more efficient manner.
+
 ![General scheme of DDROM methods [@RMP_2024].\label{fig:darom}](../images/tie_frighter.pdf){ width=100% }
 
-The techniques implemented here follow the same underlying idea expressed in \autoref{fig:darom}. They all share the typical offline/online paradigm of ROM techniques: the former is computationally expensive and it is performed only once, whereas the latter is cheap from the computational point of view and allows to have quick and reliable evaluations of the state of the system by merging background model knowledge and real evaluations of quantities of interest [@maday_parameterized-background_2014].
[…]
