### 2024-09-03 ERMETE-Lab/ROSE-pyforce#10 @Steriva
# Response Letter: Software Paper

Thank you very much for the useful comments and suggestions. Here are listed the response and the changes made to the joss-paper.

The changes are reported in pull request #12.

1. *I think the first paragraph of the next section ("Statement of need") should be switched as the "Summary" of the paper as it better describes the high-level functionality and purpose of the package. Furthermore, the summary should also include the target audience of the package which the authors have also stated in the last paragraph of the next section.*
   - [x] The paragraphs have been switched and more insights about the target audience of the package are added.
2. *"...more hostile environment compared to...": Perhaps simply "than" is better than "compared to".*
   - [x] Fixed
3. *"...and able to combine together different sources...": I think "together" is unnecessary."*
   - [x] Fixed
4. *"...and real data (i.e., measurements)...": Perhaps to be succinct, this can be written simply as "measurement data"*
   - [x] Changed to measurements data (i.e., local evaluations of quantities of interest).
5. *"...the dynamics of un-observable complex field...": Somewhere else in the paper "non-observable" is used, perhaps simply pick one?*
   - [x] Fixed.

1. *Because pyforce is an abbreviation perhaps the expanded term should appear directly after "pyforce"*
   - [x] Changed
2. *"...mainly for the Nuclear Engineering world.": Perhaps "in" instead of "for"*
   - [x] Fixed
3. *"...Nuclear Engineering...": Is the capitalization really necessary?*
   - [x] No, it isn't: changed
4. *"The package is part of the ROSE...": Is there a reference to ROSE perhaps a link or a document?*
   - [x] Some comments about ROSE (an acronym used within the ERMETE-Lab to refer to ROM related applications to multi-physics for nuclear reactors).
5. *"at reducing..., at searching... and at integrating...": I think there is no need to repeat the preposition "at" due to already parallel structure*
   - [x] Fixed
6. *Figure 1 should be better clarified by the second paragraph in this section; the figure contains mathematical symbols and notations that should be defined briefly and explicitly in the passage (if these symbols and notations are kept in the end). For instance, reading the paragraph it is not clear the starting point of the whole scheme (is it the field according to the full order model, or something else?).*
   - [x] The description of the Figure has been greatly expanded, the Figure itself has been updated to reflect the one present on the README of the Github repo.
7. *I would suggest that the list of implemented features are rewritten (perhaps in tabular form) to correspond to the figure better; which ones belong specifically to the offline and online phase, just the offline phase, or just the online phase. The list structure gives the impression that these features serve similar exchangeable purposes.*
   - [x] Implemented
8. *"...either regularised with Tikhonov or not...": Perhaps it would be better to write as "either with or without Tikhonov's regularisation"*
   - [x] Changed
9. *"...for the Online Phase": Before there was no capitalization for "Online Phase". Is the capitalization intentional?*
   - [x] No, it isn't: changed

## Contribution and authorship

[CRediT](https://credit.niso.org/) taxonomy has been added to clarify the roles, reported below.

- Stefano Riva: Conceptualization, Data curation, Formal analysis, Software, Visualization, Writing – original draft
- Carolina Introini: Conceptualization, Formal analysis, Software, Supervision, Writing – review & editing
- Antonio Cammi: Conceptualization, Project administration, Resources, Supervision, Writing – review & editing
