### 2024-09-09 4f4d748
Update classes to CamelCase convention
