### 2024-09-10 ERMETE-Lab/ROSE-pyforce#8 @damar-wicaksono
## Update (10.09.2024)

- [x] Community Guidelines

Currently, there is no section in the documentation regarding the contribution guidelines. As this is one of the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), I would ask the authors to consider adding them following the common best-practices. Per the reviewer guideline, it should include:

- Contribute to the software
- Report issues or problems with the software
- Seek support
