### 2024-09-10 ERMETE-Lab/ROSE-pyforce#10 @damar-wicaksono
@Steriva: Thank you very much for taking the time going through my comments! So that I can verify the changes and close this issue, is the updated draft can already be built into a PDF file from the [JOSS review page](https://github.com/openjournals/joss-reviews/issues/6950)? I'm asking because it seems that the pull request is still open...
