### 2024-09-13 1894835
Modification of the introduction, adding the section on theory and package structure, new images
