+++ docs/theory.md
@@ -1,0 +1,61 @@
+# Theory and package structure
+This section presents the main ideas behind Reduced Order Modelling (ROM) {cite:p}`Quarteroni2016, MadayChapter2020, Degen2020_conference`, focusing on data-driven paradigm of these techniques. Then, the structure of the package is presented showing how the different classes are connected to each other.
+
+## What is Reduced Order Modelling?
+In scientific literature, the expression Reduced Order Modelling is related to a set of techniques devoted to the search for an optimal coordinate system onto which some parametric solutions of Partial Differential Equations (PDEs) - typically called High-Fidelity (HF) or Full Order Model (FOM) - can be represented. These methods are very useful in multi-query and real-time scenarios, when quick and efficient solutions of models are required, e.g. optimization, uncertainty quantification and inverse problems {cite:p}`Guo_Veroy2021, Degen2022`. Recently, with the developments in data-driven modelling, a lot of interest in the combination of data and models has been raised. ROM offers new opportunities both to integrate the model with experimental data in real-time and to define methods of sensor positioning, by providing efficient tools to compress the prior knowledge about the system coming from the parametrized mathematical model into low-dimensional forms.
+
+### Reduced Basis Methods
+Among all ROM methods, Reduced Basis (RB) methods {cite:p}`Quarteroni2014, Hesthaven2016, Degen2020_certifiedRB` are a well-established and widely used class of ROM techniques, which are based on an offline-online paradigm. In the offline stage, a set of RB functions {math}`\phi_n(\mathbf{x})` is derived from an ensemble of high-fidelity solutions, called *snapshots*, yielding a low dimensional space that retains the main features of the full-order model. Different approaches can be used to construct the reduced basis, such as the **greedy** algorithms {cite:p}`Maday2006` and the Proper Orthogonal Decomposition **POD** {cite:p}`Berkooz1993`, directly related to the Singular Value Decomposition (SVD) {cite:p}`DataDrivenScience_book`. Regardless of the construction strategy, an approximation of the high-fidelity solution is sought during the online stage as a linear combination of the RB functions $\{\psi_n(\mathbf{x})\}$, i.e.
+
+```{math}
+    u(\mathbf{x}\;\mu) \simeq \sum_{n=1}^N\alpha_n(\mu)\cdot \psi_n(\mathbf{x})
+```
+
+According to the methodology for calculating the expansion coefficients {math}`\alpha_n(\mu)` (also called reduced coefficients or latent dynamics) of the approximation, RB methods are classified into two categories: *intrusive* and *non-intrusive* RB methods.
+
+- *Intrusive*: the governing equations of the physical system, to which the snapshots are solution, must be known and used during the online step. From a set of PDEs a rather small system of ODEs is derived, typically using Galerkin projection.
+- *Non-Intrusive*: the governing equations knowledge is not required, a more data-driven approach is followed.
+
+*pyforce* mainly focuses on the latter, since they are more suited for extension to include as input real experimental data, such as local measurements.
+
+## Data-Driven ROM techniques
+
+Data-Driven Reduced Order Modelling (DDROM) {cite:p}`RMP_2024, DDMOR_CFR` is a set of techniques, combining theoretical modelling with real data collecting from a physical system. In particular, ROM is seen in a Data Assimilation (DA) framework {cite:p}`DataDrivenScience_book`, so that the theoretical prediction, approximated by ROM, is corrected or updated by experimental evaluations of some fields (e.g., the local measurements of the temperature in a pipe or the neutron flux in a nuclear reactor).
+
[…]
