+++ README.md
@@ -18,7 +18,11 @@
   - [Recent works with *pyforce*](#recent-works-with-pyforce)
 - [Installation](#installation)
 - [Tutorials](#tutorials)
-- [Authors](#authors)
+- [Authors and contributions](#authors-and-contributions)
+- [Community Guidelines](#community-guidelines)
+  - [Contribute to the Software](#contribute-to-the-software)
+  - [Reporting Issues or Problems](#reporting-issues-or-problems)
+  - [Seeking Support](#seeking-support)
 
 ## Description
 
@@ -115,7 +119,7 @@ The *pyforce* package is tested on some tutorials available in the [docs](https:
 
 The snapshots can be either generated by the user or be downloaded at the following link [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.11483677.svg)](https://doi.org/10.5281/zenodo.11483677)
 
-## Authors
+## Authors and contributions
 
 **pyforce** is currently developed and mantained at [Nuclear Reactors Group - ERMETE Lab](https://github.com/ERMETE-Lab) by
 
@@ -125,3 +129,24 @@ The snapshots can be either generated by the user or be downloaded at the follow
 under the supervision of Prof. Antonio Cammi.
 
 If interested, please contact stefano.riva@polimi.it, carolina.introini@polimi.it, antonio.cammi@polimi.it
+
+## Community Guidelines
+
+We welcome contributions and feedback from the community! Below are the guidelines on how to get involved:
+
+### Contribute to the Software
+If you would like to contribute, please follow these steps:
+
+1. Fork the repository.
+2. Implement your changes. If you're adding new features, we kindly ask that you include an example demonstrating how to use them.
+3. Submit a pull request for review.
+
+### Reporting Issues or Problems
+If you encounter any issues or bugs with *pyforce*, please report them through the GitHub [Issues](https://github.com/ERMETE-Lab/ROSE-pyforce/issues) page. Be sure to include detailed information to help us resolve the problem efficiently.
+
+### Seeking Support
+For support, you can either:
+- Open a discussion on the GitHub [Discussions](https://github.com/ERMETE-Lab/ROSE-pyforce/discussions) page.
+- Send an email directly to: [stefano.riva@polimi.it](mailto:stefano.riva@polimi.it) or [carolina.introini@polimi.it](mailto:carolina.introini@polimi.it)
+
+Thank you for helping improve **pyforce**!
