### 2024-09-13 ERMETE-Lab/ROSE-pyforce#10 @damar-wicaksono
Thank you again for revising the paper! I went through the latest draft and confirmed that my previous remarks have been addressed.

There are a couple additional things that you might want to consider before I can close this issue.

- [x] l. 33 "_...whereas the latter is cheap from the computational point of view..._": Why not use the previous construction of "computationally expensive" and state simply "computational cheap"?
- [x] l. 34 "_...point of view and allows to have quick and reliable evaluations..._": If I'm not wrong "allow to" needs a direct object so either "allows one to have quick" or change the preposition into "allows for quick and reliable evaluations..." or change the verb to, say, "enable".
- [x] l. 36 "_During the offline..._": Perhaps start a new paragraph here to have a clear separation from what came before (general idea of offline/online phase) and what is coming after (the online phase).
- [x] l. 40 "_...reduced representation through a set of basis functions_ $`\{ \psi_n(\boldsymbol{x}) \}`$...": maybe add directly after the set "of size $N$", i.e., "_...through a set of basis functions_ $`\{ \psi_n(\boldsymbol{x}) \}`$ _of size_ $N$..."
- [x] l. 41 typo "decresed"
- [x] l. 41, similar comment as above regarding "allow to" without a direct object
- [x] l. 44 "_...allows for the search of the optimal positions of sensors...in a more efficient manner_": I think "allows for" is correct but  it should be "search for" instead of "search of" in the present usage. So perhaps: "_...allows for efficiently searching for optimal sensor placements in the physical domain._"?
- [x] l. 46 "_All these steps are performed during the offline phase,..._": In my opinion this is not necessary, simply start with this new paragraph the online phase.
- [x] l. 46 "_..., the online phase aim consists in obtaining..._": Why not "the online phase aims to obtain..."
- [x] l. 48-49 "_The DDROM online takes place which produces a novel set of reduced variables, $`\alpha^*`$, and then computing an improved reconstructed state $`\hat{u}_{\mathrm{DDROM}}`$ through a decoding step from the low dimensional state to the high dimensional one._": I find this hard to read. Perhaps as an alternative:

> "The DDROM online phase produces a new set of reduced variables (_are these "coefficients", as in the above, or "variables"?_), $`\alpha^*`$, and then  computes an improved reconstructed state $`\hat{u}_{\mathrm{DDROM}}`$ through a decoding step that transforms the low-dimensional representation to the high-dimensional one."

And with that, I have no further comments regarding on the draft manuscript. Please let me know if you need clarifications. Thank you!
