### 2024-09-15 ERMETE-Lab/ROSE-pyforce#10 @Steriva
Thanks for the comments: all the changes have been implemented and merged into the main branch with pull request #13.

The draft paper has been generated with [this action](https://github.com/ERMETE-Lab/ROSE-pyforce/actions/runs/10873140557).
