### 2024-09-16 ERMETE-Lab/ROSE-pyforce#10 @damar-wicaksono
Thank you very much for the revision! I confirm that the points I've raised have been resolved.

One additional very minor issue that you might want to correct is line 56 on page 3: "_...and its scope of application is not only restricted to the **Nuclear Engineering** world..._" which to be consistent with the previous usage should not be capitalized.

I'm closing this issue now.... Thank you!
