### 2024-09-16 489d79b
Update synthetic test error method for GEIM, PODx, TR-GEIM and PBDW
