### 2024-09-16 d67a1d5
Update self.norm in all classes for consistency
