### 2024-09-17 ERMETE-Lab/ROSE-pyforce#9 @damar-wicaksono
Thank you for going through my comments!

Here are my replies:

> The modifications these comments require will be addressed in the pull request, following issue https://github.com/ERMETE-Lab/ROSE-pyforce/issues/8.

Thank you!

> The function computeLebesgue has been split in order to execute it even after the training of the GEIM algorithm on separate codes, avoiding re-training the GEIM algorithm. This can be useful when different Lebesgue constants have to be compared.

Okay!

> The output of each method has been changed to namedtuple, as suggested, for errors and computational time to help the access to relevant quantities

Thanks!

> synt_ stands for synthetic, since the measurements are coming from simulation data (not actually collected on an experimental facility).

Thank you for the explanation. I think it would be nice to put your explanation (or a version of it) in the codebase as in-code documentation at a point(s) these terms appear prominently or in the documentation itself regarding the difference between these two terminologies ("synthetic measurements" vs "real measurements").

> We would prefer to keep this terminology to avoid importing like from pyforce.online.geim import geimOnline which we think can be a bit repetitive. I am adding some figures in the docs to help the user show how the classes are connected, as you suggested.

Understood!

> Sorry about the typo about GramSchmidt: it has been fixed.

Great!

> An overall check on the classes has been performed to be consistent with the camel case convention.

Thank you!

> Schemes of the different classes and algorithms for the offline and online phase have been added to the documentation ([link to images](https://github.com/Steriva/ROSE-pyforce/tree/8-docs-review/docs/images) and [development branch](https://github.com/Steriva/ROSE-pyforce/tree/8-docs-review)) and the read me:

I think the diagrams for classes and the required data and how they relate to each other would be a nice addition to the documentation. Thanks!

> The POD, GEIM and PBDW classes have been implemented in different time periods and to stay simple no parent/abstract class was coded; even though it is true they do share a lot of similarities, especially in the online phase. Discussing with @Neko-tan, we think it will be for sure one of the main enhancements of future releases of the package and it is planned in the future.

I understand. The current structure is not a breaking point but it would be worthwhile to revisit it in the future. Thank you.

> pyforce.offline.sensors is a class required to generate the Riesz representation of a set of sensors, either in L2 or H1, necessary by GEIM and PBDW. As already mentioned, the connection between the different classes is being added to the documentation and they should provide a clear overview of the package functionalities (for instance, the basis functions from either POD or GEIM can be mixed together with GEIM sensors in the PBDW class; other combinations are not suggested).

Yes, I think the documentation update on this would certainly be helpful.

---

I think you've addressed all my comments and I have no further points to raise in this area, except for a minor point above regarding updating the documentation a bit.

As the remaining points are deferred to the documentation, I will have a look into it separately.

I'm closing this issue now...
