### 2024-09-17 3e7e53a
Minor fix to theory section
