### 2024-09-17 1226ce5
Update plotting function
