+++ pyforce/pyforce/tools/plotting.py
@@ -1,7 +1,7 @@
 # Basic plotting tools using pyvista
 # Author: Stefano Riva, PhD Student, NRG, Politecnico di Milano
-# Latest Code Update: 31 August 2023
-# Latest Doc  Update: 31 August 2023
+# Latest Code Update: 16 September 2024
+# Latest Doc  Update: 16 September 2024
 
 import dolfinx.plot
 from dolfinx.fem import Function
@@ -10,33 +10,65 @@ import warnings
 from matplotlib import cm
 import pyvista as pv
 
-# pv.set_jupyter_backend("trame") # deprecated
-
-def get_scalar_grid(fun: Function, varname: str, real = True):
+def grids(fun: dolfinx.fem.Function, varname='u', log_plot: bool = False, 
+          mag_plot: bool = True, **kwargs):
     """
-    This function extracts the dofs of a scalar function for the plot using pyvista.
-
+    Creates a PyVista grid from a dolfinx.fem.Function and returns the warped or glyph representation.
+    
     Parameters
     ----------
-    fun : Function 
-        Function from which the dofs are extracted.
+    fun : `dolfinx.fem.Function` 
+        The function representing the field to be visualized.
     varname : str
-        Name of the variable.
-    real : boolean, optional (Default = True) 
-        Real dofs are considered, if `False` imaginary dofs are used.
+        The name to assign to the data in the PyVista grid (default is 'u').
+    log_plot : bool
+        If True, apply a logarithmic plot to scalar data (default is False).
+    mag_plot : bool
+        If True, creates a vector warp of the grid. If False, uses glyphs (default is True).
+    **kwargs
+        Additional keyword arguments passed to the function.
+    
+    Returns
+    ----------
+    grid : `pyvista.UnstructuredGrid`
+        The resulting PyVista grid, which can be visualized using PyVista plotting functions.
+    
     """
+    
+    # Create a VTK mesh from the function's function space
     topology, cells, geometry = dolfinx.plot.create_vtk_mesh(fun.function_space)
-    u_grid = pv.UnstructuredGrid(topology, cells, geometry)
+    grid = pv.UnstructuredGrid(topology, cells, geometry)
+    
+    # Handle vector fields (multiple subspaces)
+    if fun.function_space.num_sub_spaces > 0:
+        # Initialize the values array with zeros and assign the real part of the function's array
+        values = np.zeros((geometry.shape[0], 3))
+        values[:, :len(fun)] = np.real(fun.x.array.reshape(geometry.shape[0], len(fun)))
+        grid[varname] = values
 
-    if real:
-        u_grid.point_data[varname] = fun.x.array[:].real
-    else: 
-        u_grid.point_data[varname] = fun.x.array[:].imag
+        # Choose between warping by vector or using glyphs
+        if mag_plot:
+            warped = grid.warp_by_vector(varname, **kwargs)  # Apply `kwargs`
+        else:
+            warped = grid.glyph(varname, **kwargs)  # Apply `kwargs`
+        
+        return warped, values
+    
+    # Handle scalar fields (single subspace)
+    else:
+        if log_plot:
+            values = np.log10(np.real(fun.x.array[:]))    
+        else:
+            values = np.real(fun.x.array[:])
+            
+        grid.point_data[varname] = values
+        grid.set_active_scalars(varname)
+        
+        return grid, values.reshape(-1,1)
 
-    return u_grid
-
-def PlotScalar(fun: Function, filename: str, format: str = 'png', varname: str = None,
-               clim = None, colormap = cm.jet, resolution = [1080, 720]):
[…]
