### 2024-09-17 7bd897c
Update tutorial 01 and its snapshot generation
