### 2024-09-18 f9d4b93
Update plotting for tutorial 1
