+++ docs/Tutorials/01_LaminarNS/contour_plotting.py
@@ -1,0 +1,181 @@
+import vtk
+import pyvista as pv
+import dolfinx
+from matplotlib import cm
+import numpy as np
+
+from pyforce.tools.functions_list import FunctionsList
+from pyforce.tools.plotting import grids
+
+def plot_FOM_vs_ROM( fom: FunctionsList, rom: dict, mu: int, title: str, varname: str,
+                    clim = None, mag_plot = True, colormap = cm.jet, 
+                    colormap_res = cm.plasma_r, clim_res = None,
+                    filename = None, 
+                    factor : float = 0.05, tolerance : float = 0.01): 
+    
+    keys = list(rom.keys())
+    
+    resolution = [1800, 400 * ( len(keys) + 1)]
+    plotter = pv.Plotter(shape=(len(keys)+1, 2), off_screen=False, border=False, window_size=resolution)
+    
+    lab_fontsize = 25
+    title_fontsize = 35
+    zoom = 2.25
+    
+    dict_cb = dict(title = varname, width = 0.75, height = 0.2,
+                    title_font_size=title_fontsize,
+                    label_font_size=lab_fontsize,
+                    n_labels=3,
+                    color = 'k',
+                    position_x=0.125, position_y=0.745,
+                    shadow=False) 
+    
+    if fom.fun_space.num_sub_spaces > 0:
+        vector_plot = True
+    else:
+        vector_plot = False
+    
+    ############################ FOMs ###################################
+
+    plotter.subplot(0,0)
+    
+    if vector_plot:
+        if mag_plot:
+            warped_fom, values_fom = grids(fom.map(mu), mag_plot=mag_plot, varname=varname)
+        else:
+            warped_fom, values_fom = grids(fom.map(mu), mag_plot=True, varname=varname)
+            glyph_fom, _ = grids(fom.map(mu), mag_plot=False, varname=varname, factor=factor, tolerance=tolerance)
+        
+        if clim is None:
+            clim = [0., max(np.sqrt(values_fom[:, 0]**2+values_fom[:, 1]**2+values_fom[:, 2]**2))]
+    else:
+        warped_fom, _ = grids(fom.map(mu))
+        if clim is None:
+            clim = [min(fom(mu)), max(fom(mu))]
+        
+    dict_cb['title'] = 'FOM - $'+varname+'$'
+    plotter.add_mesh(warped_fom, clim = clim, cmap = colormap, show_edges=False, scalar_bar_args=dict_cb)
+    
+    if vector_plot and mag_plot == False:
+        plotter.add_mesh(glyph_fom, color='k', opacity=0.75)
+                
+    plotter.view_xy()
+    plotter.camera.zoom(zoom)
+
+    ############################ ROMs ###################################
+
+
+    for key_i in range(len(keys)):
+        plotter.subplot(1+key_i,0)
+        
+        if vector_plot:
+            if mag_plot:
+                warped_rom, _ = grids(rom[keys[key_i]].map(mu), mag_plot=mag_plot, varname=varname)
+            else:
+                warped_rom, _ = grids(rom[keys[key_i]].map(mu), mag_plot=True, varname=varname)
+                glyph_rom,  _ = grids(rom[keys[key_i]].map(mu), mag_plot=False, varname=varname, factor=factor, tolerance=tolerance)
+            
+        else:
+            warped_rom, _ = grids(rom[keys[key_i]].map(mu))
+        
+        dict_cb['title'] = keys[key_i]+' - $'+varname+'$'
+        plotter.add_mesh(warped_rom, clim = clim, cmap = colormap, show_edges=False, scalar_bar_args=dict_cb)
+        
+        if vector_plot and mag_plot == False:
+                plotter.add_mesh(glyph_rom, color='k', opacity=0.75)
+                
+        plotter.view_xy()
[…]
