### 2024-09-23 ERMETE-Lab/ROSE-pyforce#8 @damar-wicaksono
Hi!

Thank you very much for going through my comments and made the necessary and applicable revisions.

I went through the revised documentation and confirmed that most major points I raised above have been resolved.
I really appreciate the addition of the new section on the "Theory and package structure" and with the extra figures, it provides a helpful bridge between the theory of DDROM and how things are implemented in pyforce along with its convention (adopted terms, etc.).
Having this new section and the revisions made, I find following the tutorial again becomes easier and feels more consistent.

I would raise a couple of additional minor points that you might want to double check:

## [Landing page](https://ermete-lab.github.io/ROSE-pyforce/intro.html)

- [x] "_The package is part of the ROSE..._": there is no space before "which" so the term ROSE is not rendered properly.
- [x] "_...not only restricted to the Nuclear Engineering world._": "Nuclear Engineerin" is still capitalized (the one that came before was not).

## [Reduced Basis Methods](https://ermete-lab.github.io/ROSE-pyforce/theory.html#reduced-basis-methods)

- [x] "_...called *snapshots*, yielding a low dimensional space..._": missing a hyphen in "low dimensional".
- [x] When denoting RB functions are they $\phi_n(\mathbf{x})$ or $\{ \psi_n (\mathbf{x}) \}$? Both seem to appear.
- [x] The equation is missing a semicolon between $\mathbf{x}$ and $\mu$.

## [Data-Driven ROM Techniques](https://ermete-lab.github.io/ROSE-pyforce/theory.html#data-driven-rom-techniques)

- [x] Figure reference `\autoref{fig:darom}` is not resolved.
- [x] "_...reduced representation allows for the search of the optimal positions..._": I think, search "for" instead of "of".

## [Package structure](https://ermete-lab.github.io/ROSE-pyforce/theory.html#package-structure)

- [x] "_In the following, some figures sketching how the different classes are connected to each other during the offline and online phases._":  This may not be a complete sentence because it is missing a verb, either some figures "sketch", "are sketching", or "are given" at the end of the sentence.
- [x] I'm not so sure, but perhaps you can double check the use of repeated use of "measures" in this section, I've got the feeling they should have been "measurement(s)". If it's "measurement", then double check the term used in the class figures again.
- [x] "_therefore, supposing to have two coupled fields to reconstruct $(\phi, \mathbf{u})$ and only local evaluations of $\phi$ are available, two different problems arise:_": For direct state estimation, how does information of $\mathbf{u}$ is necessary to reconstruct the spatial distribution of $\phi$? Because reading the current description it seems that if you want to know about $\phi$, just do measurement on $\phi$.
- [x] "_...supposing to know the characteristic (unseen) parameter..._": Perhaps it could be better with "Assuming that the characteristic (unseen) parameter is known,..."

## [Tutorials](https://ermete-lab.github.io/ROSE-pyforce/tutorials.html)

- [x] "_... alternatively the data can be download from..._": "downloaded" instead of "download".

## [Unsteady Laminar Navier Stokes - DFG2 benchmark](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_DFG2_benchmark.html)

- [x] "_It’s 2D version..._": "It's **a** 2D version".

## [Generation of time-dependent snapshots](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_LaminarNS/01_generateFlowData.html#Generation-of-time-dependent-snapshots)

- [x] "_The implementation of these integrals is done in the `ns.py` file within the `drag_lift` class..._": Still `drag_lift` although in the code it's already `DragLift`.
- [x] "_the same directory of the benchamrk data_": Typo "benchamrk".

## [Offline Phase](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_LaminarNS/02_offline_POD.html#Offline-Phase)

- [x] "...have been download from Zenodo": "downloaded" instead of "download".
- [x] "indices" instead of "indeces".
- [x] "_At first, the correlation matrix..._": "at first" may not be appropriate here.

## [Decay of the POD eigenvalues](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_LaminarNS/02_offline_POD.html#Decay-of-the-POD-eigenvalues)

- [x] "_Here there is the first use of the offline `POD` class..._": It sounds rather awkward to me different than the rest of the tutorial... maybe "Below we use offline POD class..." or something like that.
- [x] "The option `verbose` can be actived to either show or not the progress bar.`: "activated" instead of "actived".
- [x] There's a dangling "ù" after explaining `verbose`.

## [Definition of the modes](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_LaminarNS/02_offline_POD.html#Definition-of-the-modes)

- [x] "_...using the formula of point 3._": Maybe "step 3 above" instead of "point 3".
- [x] Is it `compute_modes` (in the text) or `compute_basis` (in the codes).
- [x] Is it `normalize` (in the text` or `normalise` (in the codes).

## [Online Phase](https://ermete-lab.github.io/ROSE-pyforce/Tutorials/01_LaminarNS/03_online_POD-I.html#Online-Phase)

- [x] "_...acting as best-case scenario..._": "as the best-case scenario".
- [x] "_This produdure..._": Typo.
- [x] "_...this is a dictornary..._": Typo.

---

- [ ] Please make sure that the other two tutorials are consistent with the revisions.
