### 2024-09-24 ERMETE-Lab/ROSE-pyforce#8 @Steriva
The documentation has been reviewed following your comments in #16.
Here's the response letter.

## Landing Page

Fixed the minor issues.

## Theory

### RB methods

Fixed the minor issues. Only $\psi$ is now used to indicate the basis functions for consistency.

### Data-Driven ROM methods

Fixed the minor issues.

### Package Structure

1. *"In the following, some figures sketching how the different classes are connected to each other during the offline and online phases.": This may not be a complete sentence because it is missing a verb, either some figures "sketch", "are sketching", or "are given" at the end of the sentence.*: added
2. *I'm not so sure, but perhaps you can double check the use of repeated use of "measures" in this section, I've got the feeling they should have been "measurement(s)". If it's "measurement", then double check the term used in the class figures again.*: yes, you are right, in some phrases "measurement" is more appropriate than "measures". I have changed it accordingly.
3. *"therefore, supposing to have two coupled fields to reconstruct $(\phi, \mathbf{u})$ and only local evaluations of $\phi$ are available, two different problems arise:": For direct state estimation, how does information of u is necessary to reconstruct the spatial distribution of ϕ? Because reading the current description it seems that if you want to know about ϕ, just do measurement on ϕ.*: yes, this is true. When measures of certain quantity are available, the reconstruction follows algorithms with the direct state estimation approach; problems arise when measures of certain quantities are not available. I have changed the sentence accordingly. I add a small comment, hope it is more clear.
4. *"...supposing to know the characteristic (unseen) parameter...": Perhaps it could be better with "Assuming that the characteristic (unseen) parameter is known,..."*

## Tutorials

Fixed the minor issue.

### Unsteady Laminar Navier Stokes - DFG2 benchmark

Fixed the minor issue.

### 01 - Generate snaps

- *"The implementation of these integrals is done in the ns.py file within the drag_lift class...": Still drag_lift although in the code it's already DragLift.*: fixed.
- *"the same directory of the benchamrk data": Typo "benchamrk".*: fixed.

### Offline phases

1. *...have been download from Zenodo": "downloaded" instead of "download".*: fixed.
2. *"indices" instead of "indeces"*: fixed.
3. *"At first, the correlation matrix...": "at first" may not be appropriate here.*: fixed.
4. *"Here there is the first use of the offline POD class...": It sounds rather awkward to me different than the rest of the tutorial... maybe "Below we use offline POD class..." or something like that.*: fixed.
5. *"The option verbose can be actived to either show or not the progress bar.`: "activated" instead of "actived".*: fixed.
6. *There's a dangling "ù" after explaining verbose.*: fixed.
7. *"...using the formula of point 3.": Maybe "step 3 above" instead of "point 3".*: fixed.
8. *Is it compute_modes (in the text) or compute_basis (in the codes).* and *Is it normalize in the textornormalise` (in the codes).*: sorry for the errors, fixed.

### Online phases

1. *"...acting as best-case scenario...": "as the best-case scenario".*: fixed.
2. *"This produdure...": Typo.*: fixed.
3. *"...this is a dictornary...": Typo.*: fixed.

------------------------------------------------------------------------------------------

*Please make sure that the other two tutorials are consistent with the revisions.*: updates are expected to be pushed in the coming days.
