+++ _sources/Tutorials/03_buoyant_cavity.rst.txt
@@ -1,6 +1,15 @@
 Steady Buoyant Navier-Stokes for Differentially Heated Cavity
 =============================================================
 
+**Aim of this tutorial:** learn how the indirect reconstruction algorithm (two step approach) works, combining the methods discussed in the previous tutorial. This tutorial is of particular interest for multi-physics problems with partial observations: it is legitimate to investigate the possibility that a field carries information about another field: for instance, the temperature and the velocity in a buoyancy-driven problem. Moreover, this tutorial will show how to import snapshots from OpenFOAM simulations exploiting the *fluidfoam* module.
+
+.. image:: ../images/Tutorials/PODdiagram.svg
+  :width: 700
+  :alt: IR-scheme
+  :align: center
+
+This two step method is based on the following steps `Introini et al. (2023) <https://doi.org/10.1016/j.anucene.2022.109538>`_: at first, the characteristic parameter is estimation by solving an optimisation problem whose input are the measures of the observable field; then, the reconstruction of the unobservable field is performed by using POD with Interpolation.
+
 The Differentially Heated Cavity is taken from the `ROSE-ROM4FOAM tutorial <https://ermete-lab.github.io/ROSE-ROM4FOAM/Tutorials/BuoyantCavity/problem.html>`_.
 It's 2D Cavity with velocity imposed at the left and right boundary, along with :cite:p:`Saha2018`.
 
@@ -24,7 +33,7 @@ given :math:`\Omega` as the domain.
     :maxdepth: 3
     :caption: Steps:
 
-    Generation of the snapshots <03_BuoyantCavity_OF6/01_import_OFsnaps.ipynb>
+    Import Snapshots from OF <03_BuoyantCavity_OF6/01_import_OFsnaps.ipynb>
     Offline Phase: POD <03_BuoyantCavity_OF6/02a_offline_POD.ipynb>
     Offline Phase: GEIM <03_BuoyantCavity_OF6/02b_offline_GEIM.ipynb>
     Offline Phase: generation of the maps <03_BuoyantCavity_OF6/02c_offline_maps.ipynb>
