+++ _sources/Tutorials/03_BuoyantCavity_OF6/01_import_OFsnaps.ipynb.txt
@@ -6,9 +6,12 @@
    "source": [
     "# Importing Snapshots from OpenFOAM\n",
     "\n",
-    "This notebook uses *pyforce* to import some parametric data generated with OpenFOAM-6, exploiting the *fluidfoam* library.\n",
+    "**Aim of this notebook:** learn how to import parametric snapshots from OpenFOAM (version 6 from org-version) using *fluidfoam* package.\n",
     "\n",
-    "The snapshots are related to the Buoyant Cavity problem in fluid dynamics, governed by the Navier-Stokes equations, including energy, under the Boussinesq approximation. In particular, the snapshots have been generated using the case reported in [ROSE-ROM4FOAM tutorials](https://ermete-lab.github.io/ROSE-ROM4FOAM/Tutorials/BuoyantCavity/problem.html)."
+    "The snapshots are related to the Buoyant Cavity problem in fluid dynamics, governed by the Navier-Stokes equations, including energy, under the Boussinesq approximation. In particular, the snapshots have been generated using the case reported in [ROSE-ROM4FOAM tutorials](https://ermete-lab.github.io/ROSE-ROM4FOAM/Tutorials/BuoyantCavity/problem.html).\n",
+    "\n",
+    "*Disclaimer*: the OpenFOAM snapshots are not included in the Zenodo repository for storage issues, but they can be generated using the case reported in the link above. Therefore, to generate the OpenFOAM case and later execute this notebook, you need to have OpenFOAM-v6 installed in your machine.\n",
+    "For interested readers, please contact stefano.riva@polimi.it."
    ]
   },
   {
@@ -18,7 +21,6 @@
    "outputs": [],
    "source": [
     "import numpy as np\n",
-    "import pickle\n",
     "import matplotlib.pyplot as plt\n",
     "\n",
     "from mpi4py import MPI\n",
@@ -29,16 +31,13 @@
     "\n",
     "from pyforce.tools.functions_list import FunctionsList\n",
     "from pyforce.tools.backends import LoopProgress\n",
-    "from pyforce.tools.write_read import import_OF, StoreFunctionsList as store\n",
     "\n",
     "path = './Snapshots/OpenFOAM/'\n",
     "\n",
     "importing = ['TrainSet', 'TestSet']\n",
     "\n",
     "var_names = ['p', 'T', 'U']\n",
-    "is_vector = [False, False, True]\n",
-    "# var_names = ['U']\n",
-    "# is_vector = [True]"
+    "is_vector = [False, False, True]"
    ]
   },
   {
@@ -46,12 +45,12 @@
    "metadata": {},
    "source": [
     "The snapshots are dependent on two different parameters: the Reynolds and the Richardson number, split into train and test set\n",
-    "\\begin{equation}\n",
+    "\\begin{equation*}\n",
     "\\begin{array}{cc}\n",
     "Re_{train} = [15:5:150] & Ri_{train} = [0.2:0.4:5] \\\\\n",
     "Re_{test} = [17.5:10:147.5] & Ri_{test} = [0.4:0.8:44] \n",
     "\\end{array}\n",
-    "\\end{equation}\n"
+    "\\end{equation*}"
    ]
   },
   {
@@ -177,15 +176,19 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "fun_spaces = [FunctionSpace(domain, ('Lagrange', 1)), FunctionSpace(domain, ('Lagrange', 1)), FunctionSpace(domain, ufl.VectorElement(\"CG\", domain.ufl_cell(), 1))]\n",
[…]
