+++ _sources/Tutorials/02_MGDiffusion/02c_offline_SGREEDY.ipynb.txt
@@ -7,9 +7,11 @@
    "source": [
     "# Offline Phase: SGREEDY algorithm for sensor placement\n",
     "\n",
-    "This notebook implements the following algorithm for sensor placement:\n",
+    "**Aim of the tutorial:** this notebook shows how to use the SGREEDY algorithm for sensor placement starting from a reduced basis, e.g., the POD modes.\n",
     "\n",
-    "- Stabilised GREEDY algorithm (SGREEDY)"
+    "----\n",
+    "\n",
+    "*To execute this notebook* it is necessary to have the POD modes stored in `Offline_results/BasisFunctions` folder, placed in this directory (otherwise modify `path_off` variable)."
    ]
   },
   {
@@ -23,14 +25,10 @@
     "from IPython.display import clear_output\n",
     "import pickle\n",
     "\n",
-    "import gmsh\n",
-    "from mpi4py import MPI\n",
-    "from dolfinx.io.gmshio import model_to_mesh\n",
     "from dolfinx.fem import FunctionSpace\n",
     "\n",
-    "from pyforce.tools.write_read import ImportH5, StoreFunctionsList as store\n",
+    "from pyforce.tools.write_read import ImportH5, StoreFunctionsList\n",
     "from pyforce.tools.functions_list import FunctionsList\n",
-    "from pyforce.offline.sensors import SGREEDY\n",
     "\n",
     "path_off ='./Offline_results/'"
    ]
@@ -49,27 +47,9 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "gdim = 2\n",
+    "from neutronics import create_anl11a2_mesh\n",
     "\n",
-    "model_rank = 0\n",
-    "mesh_comm = MPI.COMM_WORLD\n",
-    "\n",
-    "# Initialize the gmsh module\n",
-    "gmsh.initialize()\n",
-    "\n",
-    "# Load the .geo file\n",
-    "gmsh.merge('ANL11A2_octave.geo')\n",
-    "gmsh.model.geo.synchronize()\n",
-    "\n",
-    "# Set algorithm (adaptive = 1, Frontal-Delaunay = 6)\n",
-    "gmsh.option.setNumber(\"Mesh.Algorithm\", 6)\n",
-    "\n",
-    "gmsh.model.mesh.generate(gdim)\n",
-    "gmsh.model.mesh.optimize(\"Netgen\")\n",
-    "\n",
-    "# Domain\n",
-    "domain, ct, ft = model_to_mesh(gmsh.model, comm = mesh_comm, rank = model_rank, gdim = gdim )\n",
-    "gmsh.finalize()\n",
+    "domain, _, _ = create_anl11a2_mesh(use_msh=True, save_mesh=False)\n",
     "\n",
     "fuel1_marker    = 1\n",
     "fuel2_marker    = 2\n",
@@ -79,21 +59,14 @@
     "void_marker     = 10\n",
     "sym_marker      = 20\n",
     "\n",
-    "tdim = domain.topology.dim\n",
-    "fdim = tdim - 1\n",
-    "domain.topology.create_connectivity(fdim, tdim)\n",
-    "\n",
     "clear_output()"
    ]
   },
   {
-   "attachments": {},
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## Importing Snapshots\n",
-    "\n",
-    "The snapshots are loaded and stored into suitable data structures."
+    "Let us import the POD modes, using the `ImportH5` function from the `pyforce` package."
    ]
   },
   {
@@ -116,41 +89,6 @@
     "                 r'\\phi_2'\n",
     "                 ]\n",
     "\n",
-    "# Snapshot path\n",
-    "path_FOM = './Snapshots/'\n",
-    "\n",
[…]
