+++ pyforce/pyforce/tools/write_read.py
@@ -1,7 +1,7 @@
 # I/O tools
 # Author: Stefano Riva, PhD Student, NRG, Politecnico di Milano
-# Latest Code Update: 09 September 2024
-# Latest Doc  Update: 05 July 2024
+# Latest Code Update: 15 October 2024
+# Latest Doc  Update: 15 October 2024
 
 # from pyforce.tools.functions_list import FunctionsList
 from .functions_list import FunctionsList
@@ -292,3 +292,238 @@ class ImportOF():
                     progressBar.update(1, percentage=False)
             
         return snap_dolfinx
+
+# ## Routine for importing from OpenFOAM using pyvista and fluidfoam --- to complete
+# class ImportOF():
+#     r"""
+#     A class used to import data from OpenFOAM and interpolate them onto *dolfinx* mesh.
+
+#     The fluidfoam library (see https://github.com/fluiddyn/fluidfoam) is exploited for the import process
+#         - Supported OpenFoam Versions : 2.4.0, 4.1 to 9 (org), v1712plus to v2212plus (com)
+        
+#     Parameters
+#     ----------
+#     path : str
+#         Path of the OpenFOAM case directory.
+#     extract_dofs : boolean, optional (Default=False)
+#         If true, the dofs :math:`x,\,y,\,z` of the OpenFOAM mesh are imported.
+        
+#     """
+#     def __init__(self, path: str, mode: str = 'pyvista', extract_dofs: bool = False) -> None:
+        
+#         self.path = path
+        
+#         if mode != 'fluidfoam' and mode != 'pyvista':
+            
+#             warnings.warn('The mode selected is not supported. The default mode is set to `pyvista`.')
+#             mode = 'pyvista'
+#         self.mode = mode
+            
+#         if mode == 'fluidfoam':
+#             if extract_dofs:
+#                 x, y, z = of.readmesh(path, verbose=False)
+#                 self.of_dofs = np.array([x, y, z])
+#             else:
+#                 self.of_dofs = None
+#         else:
+#             if not os.path.exists(path+'/foam.foam'):
+#                 with open(path+'/foam.foam', 'w') as file:
+#                     file.write('')
+#             self.reader = pv.POpenFOAMReader(path+'/foam.foam')
+
+#     def import_field(self, var_name: str, vector: bool =False, verbose: bool =True):
+#         """
+#         Importing all time instances (**skipping zero folder**) from OpenFOAM directory.
+
+#         Parameters
+#         ----------
+#         var_name : str
+#             Name of the field to import.
+#         vector : boolean, (Default: False)
+#             Labelling if the field is a vector.
+#         verbose: boolean, (Default = True) 
+#             If `True`, printing is enabled
+            
+#         Returns
+#         -------
+#         field : list
+#             Imported list of functions (each element is a `numpy.ndarray`), sorted in time.
+#         time_instants : list
+#             Sorted list of time.
+#         """
+        
+#         field = list()
+#         time_instants = list()
+        
+#         if self.mode == 'fluidfoam':
+#             if verbose:
+#                 print('Importing '+var_name+' using fluidfoam')
+            
+#             for file in os.listdir(self.path):
+                
+#                 if not ((file == '0.orig') or (file == '0')):
+#                     d = os.path.join(self.path, file)
+#                     if os.path.isdir(d):
+#                             if os.path.exists(d+'/'+var_name):
+#                                 if vector:
+#                                     field.append( of.readvector(self.path, file, var_name, verbose=False).T )
+#                                 else:
+#                                     field.append( of.readscalar(self.path, file, var_name, verbose=False).reshape(-1,1) )
+#                                 time_instants.append(float(file))
+                
[…]
