### 2024-10-15 8434975
Update README.md with new published works
