### 2024-10-15 fdf2419
Update README.md with selected works using pyforce
