### 2024-10-22 ERMETE-Lab/ROSE-pyforce#8 @damar-wicaksono
Hi @Steriva,

I went through the latest changes and can confirm that the updated tutorials are consistent with the previous comments. Thank you!

Out of curiosity, what's the relation between pyforce and ROM4FOAM (I see there are many similar implemented methods albeit in a different programming language)?
