### 2024-10-22 ERMETE-Lab/ROSE-pyforce#8 @Steriva
Hi @damar-wicaksono,
thank you for all the valuable comments and feedbacks about the package :)

ROM4FOAM was the "original" code, developed during my master's thesis and before by other colleagues. It was not easy to implement everything in C++ (OpenFOAM-like) and make it able to communicate with ML libraries available for Python. Then, we switched to a Python version able to read OpenFOAM cases and interact with scikit-learn or torch. Nowadays, we use pyforce most of the time for our works.
