### 2024-10-23 849b4b5
Add more pytests and minor fixes to sensors and functionslist
