+++ pyforce/pyforce/tools/functions_list.py
@@ -221,327 +221,3 @@ def train_test_split(params: list, fun_list: FunctionsList, test_size: float = 0
         test_fun.append(test)
         
     return train_params, test_params, train_fun, test_fun
-
-#####################################################################################################################
-
-# class FunctionsList():
-#     r"""
-#     A class wrapping a list of functions. They are stored as a `list` of `np.ndarray`.
-
-#     Parameters
-#     ----------
-#     function_space : FunctionSpace
-#         Functional Space onto which the Function are defined.
-
-#     """
-#     def __init__(self, function_space: FunctionSpace) -> None:
-#         self.fun_space = function_space
-#         self._list = list()
-
-#         tmp_fun = Function(self.fun_space).copy()
-#         self.fun_shape = tmp_fun.x.array.shape[0] # this must be a int
-
-#     def __call__(self, idx) -> np.ndarray:
-#         """
-#         Defining the class as callable, returns the idx-th element of the list
-#         """
-#         return self._list[idx]
-    
-#     def __len__(self) -> int:
-#         """
-#         Defining the length of the class as the length of the stored list
-#         """
-#         return len(self._list)
-
-#     def append(self, dofs_fun: np.ndarray) -> None:
-#         """
-#         Extend the current list by adding a new function.
-#         The dolfinx.fem.Function element is stored in a list as a numpy array, to avoid problems when the number of elements becomes large.
-#         The input can be either a `np.ndarray` or a `dolfinx.fem.Function`, in the latter case it is mapped to `np.ndarray`.
-
-#         Parameters
-#         ----------
-#         dofs_fun : np.ndarray
-#             Functions to be appended.
-
-#         """
-
-#         if isinstance(dofs_fun, Function):
-#             assert dofs_fun.x.array.shape[0] == self.fun_shape, "The input function dofs_fun has "+str(dofs_fun.x.array.shape[0])+", instead of "+str(self.fun_shape)
-#             self._list.append(dofs_fun.x.array[:])
-#         else:
-#             assert dofs_fun.shape[0] == self.fun_shape, "The input function dofs_fun has "+str(dofs_fun.shape[0])+", instead of "+str(self.fun_shape)
-#             self._list.append(dofs_fun)
-
-#     def delete(self, idx: int) -> None:
-#         """
-#         Delete a single element in position `idx`.
-        
-#         Parameters
-#         ----------
-#         idx : int
-#             Integers indicating the position inside the `_list` to delete.
-#         """
-#         del self._list[idx]
-
-#     def copy(self):
-#         """
-#         Defining the copy of the `_list` of elements
-#         """
-        
-#         return self._list.copy()
-
-#     def clear(self) -> None:
-#         """Clear the storage."""
-#         self._list = list()
-
-#     def sort(self, order: list) -> None:
-#        """
-#         Sorting the list according to order - iterable of indices.
-
-
-#         Parameters
-#         ----------
-#         order : list
-#             List of indices for the sorting.
-       
-#        """
-
-#        tmp = self._list.copy()
-#        self.clear()
-
-#        assert len(tmp) == len(order), "The order vector ("+str(len(order))+") must have the same length of the list ("+str(len(tmp))+")"
-#        for ii in range(len(order)):
-#           self.append(tmp[order[ii]])
-    
-#     def return_matrix(self) -> np.ndarray:
-#         r"""
-#         Returns the list of arrays as a matrix :math:`S\in\mathbb{R}^{\mathcal{N}_h\times N_s}`.
-#         """
-        
-#         return np.asarray(self._list).T
-
[…]
