+++ tests/test_IO.py
@@ -46,3 +46,6 @@ def test_import_h5():
     
     os.remove(str(filename)+".xdmf")
     os.remove(str(filename)+".h5")
+
+
+## To Do: add OpenFOAM import test using pyvista examples

[…]
