+++ pyforce/pyforce/tools/write_read.py
@@ -4,16 +4,24 @@
 # Latest Doc  Update: 15 October 2024
 
 # from pyforce.tools.functions_list import FunctionsList
+import warnings
+
 from .functions_list import FunctionsList
 from .backends import LoopProgress
+
 from dolfinx.io import XDMFFile, gmshio
+from dolfinx.mesh import create_mesh
+import ufl
 from mpi4py import MPI
 import h5py
 from dolfinx.fem import Function, FunctionSpace
+
 import os
+import glob
 import numpy as np
 import fluidfoam as of
-import pandas as pd
+import pyvista as pv
+
 from scipy.interpolate import NearestNDInterpolator, LinearNDInterpolator
 
 def StoreFunctionsList(domain, snap: FunctionsList, var_name: str, filename: str, order = None):
@@ -122,8 +130,252 @@ def ImportH5(V: FunctionSpace, filename: str, var_name: str, verbose = False):
 
     return snap, params_np
 
+## Routine for importing from OpenFOAM using pyvista and fluidfoam
+class ReadFromOF():
+    r"""
+    A class used to import data from OpenFOAM and interpolate them onto *dolfinx* mesh.
 
-## Routine for importing from OpenFOAM using fluidfoam
+    The fluidfoam library (see https://github.com/fluiddyn/fluidfoam) is exploited for the import process
+        - Supported OpenFoam Versions : 2.4.0, 4.1 to 9 (org), v1712plus to v2212plus (com)
+        
+    Parameters
+    ----------
+    path : str
+        Path of the OpenFOAM case directory.
+    extract_dofs : boolean, optional (Default=False)
+        If true, the dofs :math:`x,\,y,\,z` of the OpenFOAM mesh are imported.
+        
+    """
+    def __init__(self, path: str, mode: str = 'pyvista', extract_dofs: bool = False) -> None:
+        
+        self.path = path
+        
+        if mode != 'fluidfoam' and mode != 'pyvista':
+            
+            warnings.warn('The mode selected is not supported. The default mode is set to `pyvista`.')
+            mode = 'pyvista'
+        self.mode = mode
+            
+        if mode == 'fluidfoam':
+            if extract_dofs:
+                x, y, z = of.readmesh(path, verbose=False)
+                self.of_dofs = np.array([x, y, z])
+            else:
+                self.of_dofs = None
+        else:
+            # Check if any file with .foam extension exists in the directory
+            foam_files = glob.glob(os.path.join(path, '*.foam'))
+            
+            if not foam_files:
+                # If no .foam file exists, create foam.foam
+                foam_file_path = os.path.join(path, 'foam.foam')
+                with open(foam_file_path, 'w') as file:
+                    file.write('')
+            else:
+                # Use the first found .foam file
+                foam_file_path = foam_files[0]
+            
+            self.reader = pv.POpenFOAMReader(foam_file_path)
+
+    def import_field(self, var_name: str, vector: bool =False, verbose: bool =True):
+        """
+        Importing all time instances (**skipping zero folder**) from OpenFOAM directory.
+
+        Parameters
+        ----------
+        var_name : str
+            Name of the field to import.
+        vector : boolean, (Default: False)
+            Labelling if the field is a vector.
+        verbose: boolean, (Default = True) 
+            If `True`, printing is enabled
+            
+        Returns
+        -------
+        field : list
+            Imported list of functions (each element is a `numpy.ndarray`), sorted in time.
+        time_instants : list
+            Sorted list of time.
+        """
+        
+        field = list()
+        time_instants = list()
+        
+        if self.mode == 'fluidfoam':
+            if verbose:
+                print('Importing '+var_name+' using fluidfoam')
+            
+            for file in os.listdir(self.path):
+                
+                if not ((file == '0.orig') or (file == '0')):
[…]
