### 2024-10-25 #6950 @damar-wicaksono
Hi @Fei-Tao,

I've completed my portion of the review. I raised several comments and suggestions regarding the package, its documentation and accompanying paper (please refer to the linked Issues in this thread). The authors have addressed these points satisfactorily.

The package implements a range of DDROM techniques that I believe will be valuable to its target audience. The documentation includes detailed tutorials that showcase the package's capabilities, particularly in the context of nuclear engineering and related fields (as advertised), making them well-suited to its intended users. The documentation is well-structured and comprehensive, and the paper is well-written.

I am pleased to support the publication of this package in JOSS.
