### 2025-01-16 ERMETE-Lab/ROSE-pyforce#20 @alberto743
Software paper: statement of need improvements
Dear authors,
In my opinion stating in the software paper for JOSS you should better elaborate this sentence
"This framework poses more challenges in the monitoring of the system itself; since placing sensors inside the reactor itself is a nearly impossible task".
The impossibility (or not) to place sensors is very dependent on the reactor concept and you should better describe the reasoning, so that the reader may better understand your goals.
Thank you.
