### 2025-01-16 #6950 @alberto743
Issue raised on "statement of need" https://github.com/ERMETE-Lab/ROSE-pyforce/issues/20
