### 2025-01-17 ERMETE-Lab/ROSE-pyforce#20 @Steriva
Greetings,

Thank you for your comment aimed at improving the paper. We have revised the mentioned paragraph as follows:

> _Innovative reactor technologies in the framework of Generation IV are usually characterised by harsher and more hostile environments than standard nuclear systems, for instance, due to the liquid nature of the fuel or the adoption of liquid salt and molten as coolant. This novel concepts poses more challenges in the monitoring of the system itself: depending on the specific design, high temperatures, corrosive environments, and high radiation fluence make in-core sensor deployment difficult. As a result, it is crucial to study innovative methods that combine different sources of information (namely mathematical models and measurement data, even when collected outside the core) in a quick, reliable, and efficient way._

If this revised version addresses the concerns raised in the comment, I will proceed to push it to the main repository.

Best regards,
Stefano
