### 2025-01-19 ERMETE-Lab/ROSE-pyforce#20 @alberto743
It is much better in this form.
A doubt that this formulation raises is whether your project is useful only for non-standard fuels in Gen-IV reactors (e.g. molten salt reactors), or whether it is applicable also for other concepts, such as liquid metal cooled reactors.
