### 2025-01-19 ERMETE-Lab/ROSE-pyforce#20 @Steriva
The algorithms implemented in this package are totally general, even beyond nuclear reactors, thus the same approaches can be applied to other engineering systems basically in the same way. I have modified the text further

> _Innovative reactor technologies in the framework of Generation IV are usually characterised by harsher and more hostile environments than standard nuclear systems, for instance, due to the liquid nature of the fuel or the adoption of liquid salt or metals as coolant. These novel concepts pose more challenges in the monitoring of the system itself: depending on the specific design, high temperatures, corrosive environments, and high radiation fluence make in-core sensor deployment difficult. As a result, it is crucial to study innovative methods that combine different sources of information (namely mathematical models and measurement data, even when collected outside the core) in a quick, reliable, and efficient way. Even though, these techniques are general and can also be useful for conventional nuclear reactors or any engineering system._
