### 2025-01-19 ERMETE-Lab/ROSE-pyforce#20 @alberto743
Fine for me. Thank you.
