### 2025-01-19 #6950 @alberto743
@Steriva Is `conda` needed for the installation? Or is it just suggested?
