### 2025-01-19 #6950 @Steriva
> [@Steriva](https://github.com/Steriva) Is `conda` needed for the installation? Or is it just suggested?

@alberto743, I would say it's strongly recommended since it's the easiest way to install [_dolfinx_](https://github.com/FEniCS/dolfinx?tab=readme-ov-file#binary).
