### 2025-01-19 ERMETE-Lab/ROSE-pyforce#20 @Steriva
Thank you for the comments, I will commit the changes and later close this issue :)
