### 2025-01-19 #6950 @alberto743
@Steriva Thank you for the answer. Does it apply to developers as well? I guess it would be useful to specify in the documentation the instructions for both users and developers. It would ease my task to verify whether I may reproduce the compilation step from scratch.
