### 2025-01-19 #6950 @Steriva
The installation procedure is the same for both users and developers, I will specify it in the installation notes file.
