### 2025-01-19 23cae3a
Update installation notes
