### 2025-01-19 ERMETE-Lab/ROSE-pyforce#20 @Steriva
This issue has been solved by issue #21
