+++ _sources/installation.md.txt
@@ -23,8 +23,15 @@ Be sure to install *gmsh* and *gmsh-api* before *dolfinx* (the package has been 
 ## Set up a conda environment for *pyforce*
 
 **Currently *pyforce* can only be obtained by directly cloning the repository (not in PyPI or conda repository).**
+**This procedure is the suggested one for both users and developers.**
 
-It is suggested to create a conda environment: at first, clone the repository
+To ensure a quicker and easier installation, it is suggested to change the *conda-solver* to `libmamba`:
+```bash
+conda install -n base conda-libmamba-solver
+conda config --set solver libmamba
+```
+
+At first, clone the repository
 ```bash
 git clone https://github.com/ERMETE-Lab/ROSE-pyforce.git
 ```
@@ -84,7 +91,8 @@ Once all the dependencies have been installed, *pyforce* can be installed using 
 ```bash
 git clone https://github.com/ROSE-Polimi/pyforce.git
 ```
-Change directory to *pyforce* and install using pip
+Change directory to *ROSE-pyforce* and install using pip
 ```bash
+cd ROSE-pyforce
 python -m pip install pyforce/
 ```
