### 2025-05-18 ebeba55
Minor changes to trgeim and IO read
