### 2025-05-18 14e708c
Update norms calculation
