+++ pyforce/pyforce/tools/backends.py
@@ -1,20 +1,263 @@
 # Fundamental tools
 # Author: Stefano Riva, PhD Student, NRG, Politecnico di Milano
-# Latest Code Update: 30 August 2023
-# Latest Doc  Update: 25 June  2023
+# Latest Code Update: 01 November 2024
+# Latest Doc  Update: 01 November 2024
 
-from importlib import metadata
 import numpy as np
 import time
 import warnings
 
-# import dolfinx
-from dolfinx.fem import (Function, FunctionSpace, assemble_scalar, assemble_vector, form, Expression)
-from ufl import grad, inner, dx, Measure
+from dolfinx.fem import (Function, FunctionSpace, assemble_scalar, assemble_vector, assemble_matrix, form, Expression, petsc)
+from ufl import grad, inner, Measure
 from ufl.domain import extract_unique_domain
-# Note that dx will be deprecated in the future, is it necessary to import the mesh in addition to V?
 
-# Class to make progress bar
+# Class to compute norms in L2, H1 and L^\infty and the inner product in L2
+class norms():
+    r"""
+        A class to compute norms and inner products. :math:`L^2` and :math:`H^1` (semi and full are implemented for both scalar and vector fields), whereas the average and the integral are available for scalar only.
+
+        Parameters
+        ----------
+        V : FunctionSpace
+            Functional Space onto which the Function are defined.
+        is_H1 : boolean, optional (Default = False)
+            If the function belongs to :math:`H^1`, the forms for the inner products and norms are computed.
+
+    """
+    def __init__(self, V: FunctionSpace, is_H1 = False, metadata_degree=4):
+        
+        self.V = V
+        self.u1 = Function(V).copy()
+        self.u2 = Function(V).copy()
+        
+        # Deprecation warning in fenics-dolfinx=0.6.0 --> try to correct this?
+        warnings.filterwarnings("ignore", category=DeprecationWarning) 
+        
+        metadata = {"quadrature_degree": metadata_degree}
+        self.dx = Measure("dx", domain=extract_unique_domain(self.u1), metadata=metadata)
+
+        # Definition of the variational forms
+        if V.num_sub_spaces == 0: # if the functional space is related to a scalar function
+            self.integ_form = form(self.u1 * self.dx) 
+
+        self.L2form_inner = form(inner(self.u1, self.u2) * self.dx)
+        
+        if is_H1:
+            self.semiH1form_inner = form(inner(grad(self.u1), grad(self.u2)) * self.dx)
+            self.fullH1form_inner = form( (inner(grad(self.u1), grad(self.u2)) + inner(self.u1, self.u2)) * self.dx)
+            
+    def check_input(self, input: Function):
+        r"""
+        Check that the input is either a Function or a numpy array of the right shape.
+        If not, the code will probably produce an error.
+
+        This method is meant to be used internally.
+        """
+
+        if isinstance(input, Function):
+            _u  = input.x.array[:]
+        else:
+            _u = input
+
+        return _u
+    
+    def integral(self, u: Function):
+        r""" 
+        Computes the integral of a given scalar function `u` over the domain
+
+        .. math::
+            \int_\Omega u \,d\Omega 
+
+        Parameters
+        ----------
+        u : `Function` (or `np.ndarray`)
+            Function belonging to the same functional space V (it must be a scalar!)
+
+        Returns
+        -------
+        value : float
+            Integral over the domain
+        """
+        
+        _u = self.check_input(u)
+        
+        self.u1.x.array[:] = _u
+            
+        value = assemble_scalar(self.integ_form)
+            
+        return value
+
+    def average(self, u: Function):
+        r""" 
[…]
