### 2025-05-24 be3b873
Add a check for the guess and clip it to the bounds if necessary for IR algorithm (two step)
