### 2025-06-16 0d4d86e
Add classes for malfunctioning sensors following SNAMC conference
