+++ pyforce/pyforce/tools/write_read.py
@@ -397,172 +397,172 @@ class ReadFromOF():
                 
         return snap_dolfinx
 
-## Routine for importing from OpenFOAM using fluidfoam - Deprecated?
-class ImportOF():
-    r"""
-    A class used to import data from OpenFOAM and interpolate them onto *dolfinx* mesh.
+## Routine for importing from OpenFOAM using fluidfoam - Deprecated
+# class ImportOF():
+#     r"""
+#     A class used to import data from OpenFOAM and interpolate them onto *dolfinx* mesh.
 
-    The fluidfoam library (see https://github.com/fluiddyn/fluidfoam) is exploited for the import process
-        - Supported OpenFoam Versions : 2.4.0, 4.1 to 9 (org), v1712plus to v2212plus (com)
+#     The fluidfoam library (see https://github.com/fluiddyn/fluidfoam) is exploited for the import process
+#         - Supported OpenFoam Versions : 2.4.0, 4.1 to 9 (org), v1712plus to v2212plus (com)
         
-    Parameters
-    ----------
-    path : str
-        Path of the OpenFOAM case directory.
-    extract_dofs : boolean, optional (Default=False)
-        If true, the dofs :math:`x,\,y,\,z` of the OpenFOAM mesh are imported.
+#     Parameters
+#     ----------
+#     path : str
+#         Path of the OpenFOAM case directory.
+#     extract_dofs : boolean, optional (Default=False)
+#         If true, the dofs :math:`x,\,y,\,z` of the OpenFOAM mesh are imported.
         
-    """
-    def __init__(self, path: str, extract_dofs = False) -> None:
+#     """
+#     def __init__(self, path: str, extract_dofs = False) -> None:
         
-        self.path = path
+#         self.path = path
         
-        if extract_dofs:
-            x, y, z = of.readmesh(path, verbose=False)
-            self.of_dofs = np.array([x, y, z])
-        else:
-            self.of_dofs = None
+#         if extract_dofs:
+#             x, y, z = of.readmesh(path, verbose=False)
+#             self.of_dofs = np.array([x, y, z])
+#         else:
+#             self.of_dofs = None
 
-    def import_field(self,  var_name: str, vector=False, verbose=True):
-        """
-        Importing all time instances (**skipping zero folder**) from OpenFOAM directory.
+#     def import_field(self,  var_name: str, vector=False, verbose=True):
+#         """
+#         Importing all time instances (**skipping zero folder**) from OpenFOAM directory.
 
-        Parameters
-        ----------
-        var_name : str
-            Name of the field to import.
-        vector : boolean, (Default: False)
-            Labelling if the field is a vector.
-        verbose: boolean, (Default = True) 
-            If `True`, printing is enabled
+#         Parameters
+#         ----------
+#         var_name : str
+#             Name of the field to import.
+#         vector : boolean, (Default: False)
+#             Labelling if the field is a vector.
+#         verbose: boolean, (Default = True) 
+#             If `True`, printing is enabled
             
-        Returns
-        -------
-        field : list
-            Imported list of functions (each element is a `numpy.ndarray`), sorted in time.
-        time_instants : list
-            Sorted list of time.
-        """
+#         Returns
+#         -------
+#         field : list
+#             Imported list of functions (each element is a `numpy.ndarray`), sorted in time.
+#         time_instants : list
+#             Sorted list of time.
+#         """
         
-        if verbose:
-            print('Importing '+var_name)
+#         if verbose:
+#             print('Importing '+var_name)
         
-        field = list()
-        time_instants = list()
+#         field = list()
+#         time_instants = list()
         
-        for file in os.listdir(self.path):
[…]
