+++ README.md
@@ -15,7 +15,7 @@
 
 - [Description](#description)
 - [How to cite *pyforce*](#how-to-cite-pyforce)
-  - [Recent works with *pyforce*](#recent-works-with-pyforce)
+  - [Selected works with *pyforce*](#selected-works-with-pyforce)
 - [Installation](#installation)
 - [Tutorials](#tutorials)
 - [Authors and contributions](#authors-and-contributions)
@@ -119,7 +119,7 @@ The *pyforce* package is tested on some tutorials available in the [docs](https:
 
 *Coming Soon*: multiphysics (neutronics+thermal-hydraulics) with *dolfinx* and OpenFOAM.
 
-The snapshots can be either generated by the user or be downloaded at the following link [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.11483677.svg)](https://doi.org/10.5281/zenodo.11483677)
+The snapshots can be either generated by the user or be downloaded at the following link [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13882851.svg)](https://zenodo.org/records/13882851)
 
 ## Authors and contributions
 
