### 2025-06-20 355c354
Add new tutorial with import from OpenFOAM with pyvista, generation of surrogate model with GPR
