+++ README.md
@@ -119,7 +119,7 @@ The *pyforce* package is tested on some tutorials available in the [docs](https:
 
 *Coming Soon*: multiphysics (neutronics+thermal-hydraulics) with *dolfinx* and OpenFOAM.
 
-The snapshots can be either generated by the user or be downloaded at the following link [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13882851.svg)](https://zenodo.org/records/13882851)
+The snapshots can be either generated by the user or be downloaded at the following link [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15705990.svg)](https://zenodo.org/records/15705990)
 
 ## Authors and contributions
 
