### 2025-12-09 #6950 @ShimingYIN
@Steriva @Neko-tan Sorry for the long delay in the review of this JOSS submission.. I have just completed my portion of the review and raised comments and revision suggestions in the github issues:

* [Issue #24](https://github.com/ERMETE-Lab/ROSE-pyforce/issues/24)
* [Issue #25](https://github.com/ERMETE-Lab/ROSE-pyforce/issues/25)

I think the coherent suites of POD, GEIM, PBDW, indirect reconstruction algorithms in this pyforce package have demonstrated the effectiveness on multiphysics systems in NE applications, and it's fitting for the ROM community. After the previous revisions based on the other reviewers’ comments, the documentation and JOSS paper are also detailed and well-organised now. If the authors could further address the remaining issues with revisions, I'd support the acceptance of this package for JOSS publication.
