### 2025-12-09 5de86d1
Minor fix of testing related to issue #24
