+++ pyforce/requirements.txt
@@ -10,3 +10,4 @@ pyvista>=0.42.2
 scikit_learn>=1.3.1
 scipy>=1.13.0
 setuptools==63.0.0
+pytest
