### 2025-12-09 b8a1381
Update docs and readme for issue #25
