+++ README.md
@@ -17,7 +17,9 @@
 - [How to cite *pyforce*](#how-to-cite-pyforce)
   - [Selected works with *pyforce*](#selected-works-with-pyforce)
 - [Installation](#installation)
+- [Package Structure](#package-structure)
 - [Tutorials](#tutorials)
+  - [Demo results](#demo-results)
 - [Authors and contributions](#authors-and-contributions)
 - [Community Guidelines](#community-guidelines)
   - [Contribute to the Software](#contribute-to-the-software)
@@ -37,10 +39,10 @@ The techniques implemented here follow the same underlying idea expressed in the
 
 At the moment, the following techniques have been implemented:
 
-- **Proper Orthogonal Decomposition** with Projection and Interpolation for the Online Phase
-- **Generalised Empirical Interpolation Method**, either regularised with Tikhonov or not
-- **Parameterised-Background Data-Weak formulation**
-- an **Indirect Reconstruction** algorithm to reconstruct non-observable fields
+- **Proper Orthogonal Decomposition** with Projection and Interpolation for the Online Phase -> [`pyforce.offline.POD`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/offline/pod.html#POD), [`pyforce.online.pod_projection`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/online/pod_projection.html#PODproject), [`pyforce.online.pod_interpolation`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/online/pod_interpolation.html#PODI)
+- **Generalised Empirical Interpolation Method**, either regularised with Tikhonov or not -> [`pyforce.offline.geim`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/offline/geim.html#GEIM), [`pyforce.online.geim`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/online/geim.html#GEIM), [`pyforce.online.tr_geim`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/online/tr_geim.html#TRGEIM)
+- **Parameterised-Background Data-Weak formulation** -> [`pyforce.online.pbdw`](https://ermete-lab.github.io/ROSE-pyforce/_modules/pyforce/online/pbdw.html#PBDW)
+- an **Indirect Reconstruction** algorithm to reconstruct non-observable fields -> [`pyforce.online.indirect_recon`](https://ermete-lab.github.io/ROSE-pyforce/api/pyforce.online.html#module-pyforce.online.indirect_recon)
 
 This package is aimed to be a valuable tool for other researchers, engineers, and data scientists working in various fields, not only restricted in the Nuclear Engineering world.
 
@@ -104,11 +106,17 @@ create a conda environment using `environment.yml`
 cd ROSE-pyforce
 conda env create -f pyforce/environment.yml
 ```
-activate the environment and then install the package using `pip`
+activate the environment and then install the package using `pip` (be aware than on PyPI there exists another package named *pyforce*, so be sure to install it from the cloned repository)
 ```bash
 conda activate pyforce-env
-python -m pip install pyforce/
+cd pyforce/
+python -m pip install .
 ```
+
+## Package Structure
+The package **pyforce** comprises 3 subpackages: *offline*, *online* and *tools*. The first two collect the main functionalities, in particular the different DDROM techniques; whereas, the last includes importing and storing functions (from *dolfinx* directly or mapping from OpenFOAM), some backend classes for the snapshots and the calculation of integrals/norms. In the following, some figures are sketching how the different classes are connected to each other during the offline and online phases.
+
[…]
