### 2025-12-09 8f9defa
update paper according to issue #25
