+++ joss-paper/paper.md
@@ -35,7 +35,7 @@ Innovative reactor technologies in the framework of Generation IV are usually ch
 
 The techniques implemented here follow the same underlying idea expressed in \autoref{fig:darom}. They all share the typical offline/online paradigm of ROM techniques: the former is computationally expensive and it is performed only once, whereas the latter is computationally cheap and allows for a quick and reliable evaluations of the state of the system by merging background model knowledge and real evaluations of quantities of interest [@maday_parameterized-background_2014].
 
-During the offline (also called training) phase, a *high-fidelity* or Full Order Model (FOM), usually parameterised partial differential equations, is solved several times to obtain a collections of snapshots $\mathbf{u}_{FOM}\in\mathbb{R}^{\mathcal{N}_h}$, given $\mathcal{N}_h$ the dimension of the spatial mesh, which are dependent on some parameters $\boldsymbol{\mu}_n$; then, these snapshots are used to generate a reduced representation through a set of basis functions $\{\psi_n(\mathbf{x})\}$ of size $N$, in this way the degrees of freedom are decreased from $\mathcal{N}_h$ to $N$, provided that $\mathcal{N}_h>>N$. This allows for an approximation of any solution of the FOM as follows
+During the offline (also called training) phase, a *high-fidelity* or Full Order Model (FOM), usually parameterised partial differential equations, is solved several times to obtain a collection of snapshots $\mathbf{u}_{FOM}\in\mathbb{R}^{\mathcal{N}_h}$, given $\mathcal{N}_h$ the dimension of the spatial mesh, which are dependent on some parameters $\boldsymbol{\mu}_n$; then, these snapshots are used to generate a reduced representation through a set of basis functions $\{\psi_n(\mathbf{x})\}$ of size $N$, in this way the degrees of freedom are decreased from $\mathcal{N}_h$ to $N$, provided that $\mathcal{N}_h>>N$. This allows for an approximation of any solution of the FOM as follows
 
 \begin{equation}\label{eq:rb}
 u(\mathbf{x};\boldsymbol{\mu}) \simeq \sum_{n=1}^N\alpha_n(\boldsymbol{\mu})\cdot \psi_n(\mathbf{x})
@@ -71,8 +71,9 @@ Up to now, the techniques, reported in the following tables, have been implement
 | Parameterised-Background Data-Weak (PBDW) [@maday_parameterized-background_2014] |                                       | X                                        |
 | Indirect Reconstruction: parameter estimation [@introini_non-intrusive_2023]     |                                       | X                                        |
 
+The package is organized in modules for the offline and online phases, with sub-modules for each of the algorithms mentioned above, in addition to modules for handling the data, the sensors, and the collections of functions/snapshots. The official documentation provides additional details on how the different modules and classes are related to each other and how to use them.
 
-This package aims to become a valuable tool for other researchers, engineers, and data scientists working in various fields where multi-physics problems play an important role, and its scope of application is not only restricted to the Nuclear Engineering world. The package also includes tutorials showing how to use the library and its main features, ranging from snapshot generation in dolfinx, import and mapping from OpenFOAM [@weller_tensorial_1998], to the offline and online phase of each of the aforementioned DDROM algorithms. The case studies are taken from the fluid dynamics and neutronics world, being the most important physics involved in nuclear reactor physics, although the methodologies can be extended to any physics of interest.
[…]
