### 2025-12-11 #6950 @ShimingYIN
@Steriva Thanks for addressing my comments and for the updates regarding the issues.

@Fei-Tao I have completed my review. I consider the documentation and paper are well-constructed now, with its tutorials clearly demonstrating the DDROM performance. This pyforce package will be valuable to the ROM community. I recommend acceptance of this JOSS submission.
