### 2026-01-08 a9ea6bd
Update DOI format in paper.bib
