+++ joss-paper/paper.bib
@@ -35,7 +35,7 @@ volume = {135},
 pages = {243-268},
 year = {2024},
 issn = {0307-904X},
-doi = {https://doi.org/10.1016/j.apm.2024.06.040},
+doi = {10.1016/j.apm.2024.06.040},
 url = {https://www.sciencedirect.com/science/article/pii/S0307904X24003196},
 author = {Stefano Riva and Carolina Introini and Antonio Cammi},
 keywords = {Reduced order modelling, Data driven, Nuclear reactors, Multi-physics, Model correction},
