### 2026-01-08 ef77c83
Modify author details and update publication date
Updated author information and affiliations in the paper metadata.
