### 2026-01-12 f6cbd1d
Minor fixes on the code IO and installation
