+++ rose-pyforce/pyproject.toml
@@ -7,8 +7,8 @@ name = "rose-pyforce"
 version = "1.0.0"
 description = "Python Framework data-driven model Order Reduction for multi-physiCs problEms"
 readme = "README.md"
-requires-python = ">=3.10"
-license = { text = "MIT" }
+requires-python = ">=3.9"
+license = { file = "LICENSE" }
 authors = [
     { name = "Stefano Riva", email = "stefano.riva@polimi.it" },
     { name = "Carolina Introini", email = "carolina.introini@polimi.it" },
@@ -20,6 +20,8 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering :: Mathematics"
 ]
 
@@ -29,16 +31,22 @@ dependencies = [
     "matplotlib",
     "pyvista[jupyter]",
     "h5py",
-    "ipykernel",
     "tqdm",
     "fluidfoam",
     "trame-vtk",
-    "scikit-learn",
-    "setuptools>=63.0.0"
+    "scikit-learn"
+]
+
+[project.optional-dependencies]
+dev = [
+    "ipykernel",
+    "pytest",
+    "pytest-cov"
 ]
 
 [project.urls]
 Homepage = "https://github.com/ROSE-Polimi/pyforce"
 
 [tool.setuptools.packages.find]
-exclude = ["tests"]
+where = ["."]
+exclude = ["tests*"]
