### 2026-01-12 758dc90
Update docs structure with changelog, contribution and migration discussion from original version
