+++ docs/contributing.md
@@ -1,0 +1,20 @@
+# Community Guidelines
+
+We welcome contributions and feedback from the community! Below are the guidelines on how to get involved:
+
+## Contribute to the Software
+If you would like to contribute, please follow these steps:
+
+1. Fork the repository.
+2. Implement your changes. If you're adding new features, we kindly ask that you include an example demonstrating how to use them.
+3. Submit a pull request for review.
+
+## Reporting Issues or Problems
+If you encounter any issues or bugs with *pyforce*, please report them through the GitHub [Issues](https://github.com/ERMETE-Lab/ROSE-pyforce/issues) page. Be sure to include detailed information to help us resolve the problem efficiently.
+
+## Seeking Support
+For support, you can either:
+- Open a discussion on the GitHub [Discussions](https://github.com/ERMETE-Lab/ROSE-pyforce/discussions) page.
+- Send an email directly to: [stefano.riva@polimi.it](mailto:stefano.riva@polimi.it) or [carolina.introini@polimi.it](mailto:carolina.introini@polimi.it)
+
+Thank you for helping improve **pyforce**!
