+++ docs/intro.md
@@ -4,13 +4,19 @@
 
 ![pyforce](images/immy_pyforce2.png)
 
-[![Reference Paper 1](https://img.shields.io/badge/Reference%20Paper%201-10.1016/j.apm.2024.06.040-gray?labelColor=blue&style=flat&link=https://doi.org/10.1016/j.apm.2024.06.040)](https://doi.org/10.1016/j.apm.2024.06.040) [![Reference Paper 2](https://img.shields.io/badge/Reference%20Paper%202-10.1016/j.nucengdes.2024.113105-gray?labelColor=blue&style=flat&link=https://www.sciencedirect.com/science/article/pii/S002954932400205X)](https://www.sciencedirect.com/science/article/pii/S002954932400205X)
+[![JOSS Paper](https://img.shields.io/badge/JOSS%20Paper-10.21105/joss.06950-brightgreen?style=flat&logo=journal-of-open-source-software&logoColor=white)](https://doi.org/10.21105/joss.06950)
+[![Reference Paper 1](https://img.shields.io/badge/Ref%20Paper%201-Nucl.%20Eng.%20Des.-gray?labelColor=blue&style=flat&link=https://doi.org/10.1016/j.nucengdes.2024.113105)](https://doi.org/10.1016/j.nucengdes.2024.113105) [![Reference Paper 2](https://img.shields.io/badge/Ref%20Paper%202-Appl.%20Math.%20Mod.-gray?labelColor=blue&style=flat&link=https://doi.org/10.1016/j.apm.2024.06.040)](https://doi.org/10.1016/j.apm.2024.06.040)
 
 [![Zenodo](https://img.shields.io/badge/Zenodo-purple?style=flat&link=https://zenodo.org/records/15705990)](https://zenodo.org/records/15705990)
 
 ## Description
 
-*pyforce* (Python Framework for data-driven model Order Reduction of multi-physiCs problEms) is a Python package implementing some Data-Driven Reduced Order Modelling (DDROM) techniques for applications to multi-physics problems, mainly set in the **nuclear engineering** world. These techniques have been implemented upon the [pyvista](https://docs.pyvista.org/index.html) package to handle meshes, integral calculation and plotting functionalities. The package is part of the **ROSE (Reduced Order modelling with data-driven techniques for multi-phySics problEms)** framework which is one of the main research topics investigated at the [ERMETE-Lab](https://github.com/ERMETE-Lab#reduced-order-modelling-with-data-driven-techniques-for-multi-physics-problems-rose-): in particular, the focus of the research activities is on mathematical algorithms aimed at reducing the complexity of multi-physics models with a focus on nuclear reactor applications, searching for optimal sensor positions and integrating experimental data to improve the knowledge on the physical systems.
+*pyforce* (**Py**thon **F**ramework for data-driven model **O**rder **R**eduction of multi-physi**C**s probl**E**ms) is a Python package implementing Data-Driven Reduced Order Modelling (DDROM) techniques for applications to multi-physics problems, mainly set in the **nuclear engineering** world.
+
+> **⚠️ Note on Versioning**
+> This documentation describes **pyforce v1.0+**, which uses `pyvista` and `numpy` for solver-independent mesh handling and integral calculations. The original implementation (described in the JOSS paper) was built upon `dolfinx`. While the architecture has evolved, the underlying methodology remains consistent.
+
[…]

[…]
