### 2026-01-12 c31b544
Update links for installation
