### 2026-01-16 #6950 @kyleniemeyer
@Steriva before publishing, please update the metadata on the Zenodo archive to match the paper. Currently, the author/contributor lists are different.
