### 2026-01-16 #6950 @Steriva
Hi @kyleniemeyer, sorry for the missing metadata. I have updated the authors list (https://zenodo.org/records/18189656).
Let me know if anything more is needed
